% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scPagwas-data.R
\docType{data}
\name{block_annotation}
\alias{block_annotation}
\title{block_annotation}
\format{
a data.frame,
}
\source{
Generated from gencode.v34.annotation.gtf.gz
}
\usage{
data(block_annotation)
}
\description{
A gene annotation files from \code{
library("rtracklayer")
gtf_df<- rtracklayer::import("/share/pub/dengcy/GWAS_Multiomics/pagwas/data/gencode.v34.annotation.gtf.gz")
gtf_df <- as.data.frame(gtf)
gtf_df <- gtf_df[,c("seqnames","start","end","type","gene_name")]
gtf_df <- gtf_df[gtf_df$type=="gene",]
block_annotation<-gtf_df[,c(1,2,3,5)]
colnames(block_annotation)<-c("chrom", "start","end","label")
save(block_annotation,file = "/share/pub/dengcy/GWAS_Multiomics/pagwas/data/block_annotation.RData")
}
}
\examples{
data(block_annotation)
str(block_annotation)
}
\keyword{datasets}
