% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scPagwas-data.R
\docType{data}
\name{block_annotation_hg37}
\alias{block_annotation_hg37}
\title{block_annotation_hg37}
\format{
a data.frame,
}
\source{
Generated from gencode.v44lift37.annotation.gff3
}
\usage{
data(block_annotation_hg37)
}
\description{
GRCh37 gencode.v44lift37.annotation.gff3
}
\examples{
data(block_annotation_hg37)
str(block_annotation_hg37)
}
\keyword{datasets}
