% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_result_visualize.R
\name{bootstrap_estimate_Plot}
\alias{bootstrap_estimate_Plot}
\title{bootstrap_estimate_Plot}
\usage{
bootstrap_estimate_Plot(
  Pagwas,
  filenames = "test_forest.pdf",
  width = 9,
  height = 7
)
}
\arguments{
\item{Pagwas}{Pagwas format of result}

\item{filenames}{the filename of the output plot, default is test_forest.pdf"}

\item{width}{figure width}

\item{height}{figure height}
}
\value{

}
\description{
bootstrap_estimate_Plot
}
\examples{
library(scPagwas)
bootstrap_estimate_Plot(Pagwas,filenames="test_forest.pdf",width = 9,height =7)

}
