% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Link_pathway_blocks_gwas.R
\name{celltypes_Pathway_block_func}
\alias{celltypes_Pathway_block_func}
\title{celltypes_Pathway_block_func}
\usage{
celltypes_Pathway_block_func(Pagwas, pa_blocksnplist)
}
\arguments{
\item{pa_blocksnplist}{}
}
\value{

}
\description{
celltypes_Pathway_block_func
}
