% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Singlecell_heritability_contributions.R
\name{Corr_Random}
\alias{Corr_Random}
\title{Corr_Random
Split the large-scale single-cell data and perform random computations of correlation. Multiple iterations of random computations ensure the accuracy of the results.}
\usage{
Corr_Random(
  data_mat,
  scPagwas.gPAS.score,
  seed = 1234,
  random = T,
  Nrandom = 5,
  Nselect = 10000
)
}
\arguments{
\item{data_mat}{the expression matrix for single cell.}

\item{scPagwas.gPAS.score}{the gPas score for scPagwas}

\item{seed}{random seed.}

\item{random}{whether to random, default is TRUE.}

\item{Nrandom}{the number of random.}

\item{Nselect}{the number of single cells for each random.}
}
\description{
Corr_Random
Split the large-scale single-cell data and perform random computations of correlation. Multiple iterations of random computations ensure the accuracy of the results.
}
