% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_CorrectBg_p.R
\name{correct_background}
\alias{correct_background}
\title{correct_background}
\usage{
correct_background(scPagwas.TRS.Score, mat_ctrl_raw_score, v_var_ratio_c2t)
}
\arguments{
\item{scPagwas.TRS.Score}{raw score}

\item{mat_ctrl_raw_score}{control raw score}

\item{v_var_ratio_c2t}{variance ratio}
}
\value{
correct_pdf
}
\description{
correct_background
}
