% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_functions.R
\name{corSparse}
\alias{corSparse}
\title{various association measures between sparse matrices}
\usage{
corSparse(X, Y)
}
\arguments{
\item{X}{matrix}

\item{Y}{matrix or vector}
}
\value{

}
\description{
Pearson correlation matrix between columns of X, Y
modification the code from:
http://stackoverflow.com/questions/5888287/running-cor-or-any-variant-over-a-sparse-matrix-in-r
https://github.com/cysouw/qlcMatrix/blob/master/R/assoc.R
}
\note{
Note that results larger than 1e4 x 1e4 will become very slow, because the resulting matrix is not sparse anymore.

covmat uses E\link{(X-muX)'(Y-muY)} = E\link{X'Y} - muX'muY
with sample correction n/(n-1) this leads to cov = ( X'Y - n*muX'muY ) / (n-1)

the sd in the case Y!=NULL uses E\link{X-mu}^2 = E\link{X^2}-mu^2
with sample correction n/(n-1) this leads to sd^2 = ( X^2 - n*mu^2 ) / (n-1)
}
