% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Celltype_heritability_contributions.R
\name{cv_regularized_parameter_estimator}
\alias{cv_regularized_parameter_estimator}
\title{cv_regularized_parameter_estimator}
\usage{
cv_regularized_parameter_estimator(vectorized_Pagwas_data, n_folds = 10, ...)
}
\arguments{
\item{vectorized_Pagwas_data}{Pagwas data used for inference}

\item{n_folds}{number of folds for cross validation}

\item{...}{other arguments to pass to cv.glmnet}
}
\value{

}
\description{
Perform regularization inference.
Use CV to find appropriate values of lambda for either feature selection
}
