% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scPagwas-data.R
\docType{data}
\name{eqtls_files}
\alias{eqtls_files}
\title{eqtls_files
Generated from GTEx}
\format{
vector
}
\source{
Generated from GTEx
}
\usage{
eqtls_files
}
\description{
eqtls_files
Generated from GTEx
}
\examples{
data(eqtls_files)
str(eqtls_files)
}
\keyword{datasets}
