% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scPagwas_Visualization.R
\name{fortify.Seurat.umap}
\alias{fortify.Seurat.umap}
\title{fortify.Seurat.umap}
\usage{
\method{fortify}{Seurat.umap}(x)
}
\arguments{
\item{x}{Seurat format}
}
\description{
set data frame to ggplot
}
