% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scPagwas-data.R
\docType{data}
\name{genes.by.gobp.pathway}
\alias{genes.by.gobp.pathway}
\title{genes.by.gobp.pathway}
\format{
list
}
\usage{
data(genes.by.gobp.pathway)
}
\description{
A go term bp  gene set from \code{
x <- readLines("c5.go.bp.v7.5.1.symbols.gmt")
res <- strsplit(x, "\t")
names(res) <- vapply(res, function(y) y[1], character(1))
genes.by.gobp.pathway <- lapply(res, "[", -c(1:2))
}
}
\examples{
data(genes.by.gobp.pathway)
str(genes.by.gobp.pathway)
}
\keyword{datasets}
