% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scPagwas-data.R
\docType{data}
\name{Genes_by_pathway.kegg}
\alias{Genes_by_pathway.kegg}
\title{KEGG pathway gene list}
\format{
list.
}
\source{
Generated from KEGGREST
}
\description{
KEGG pathway gene list
}
\details{

}
\section{Pull all genes for each pathway}{
pathway.codes <- sub("path:", "", names(pathways.list))
Genes_by_pathway.kegg <- sapply(pathway.codes,function(pwid){
pw <- keggGet(pwid)
if (is.null(pw[\link{1}]$GENE)) return(NA)
pw2 <- pw[\link{1}]$GENE\link{c(FALSE,TRUE)} # may need to modify this to c(FALSE, TRUE) for other organisms
pw2 <- unlist(lapply(strsplit(pw2, split = ";", fixed = T), function(x)x\link{1}))
return(pw2)
})
}

\examples{
data(Genes_by_pathway.kegg)
str(Genes_by_pathway.kegg)
}
