% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scPagwas-data.R
\docType{data}
\name{Genes_by_pathway_kegg}
\alias{Genes_by_pathway_kegg}
\title{Genes_by_pathway_kegg}
\format{
a list.
}
\usage{
data(Genes_by_pathway_kegg)
}
\description{
A few numbers from \code{
library(KEGGREST)
pathways.list <- keggList("pathway", "hsa")
# Pull all genes for each pathway
pathway.codes <- sub("path:", "", names(pathways.list))
Genes_by_pathway.kegg <- sapply(pathway.codes,function(pwid){
pw <- keggGet(pwid)
if (is.null(pw[[1]]$GENE)) return(NA)
pw2 <- pw[[1]]$GENE[c(FALSE,TRUE)]
# may need to modify this to c(FALSE, TRUE) for other organisms
pw2 <- unlist(lapply(strsplit(pw2, split = ";", fixed = T), function(x)x[1]))
 return(pw2)
})

}
}
\examples{
data(Genes_by_pathway_kegg)
str(Genes_by_pathway_kegg)
}
\keyword{datasets}
