% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scPagwas-data.R
\docType{data}
\name{Genes_by_pathway_reactome}
\alias{Genes_by_pathway_reactome}
\title{Genes_by_pathway_reactome}
\format{
list
}
\source{
Generated from c2.cp.reactome.v7.5.1.symbols.gmt
}
\description{
A reactome pathway gene set from \code{
x <- readLines("E:/OneDrive/GWAS_Multiomics/tempdata/c2.cp.reactome.v7.5.1.symbols.gmt")
res <- strsplit(x, "\t")
names(res) <- vapply(res, function(y) y[1], character(1))
Genes_by_pathway.reactome <- lapply(res, "[", -c(1:2))
save(Genes_by_pathway.reactome,file="E:/RPakage/scPagwas/data/Genes_by_pathway.reactome.RData")
}
}
\examples{
data(Genes_by_pathway_reactome)
str(Genes_by_pathway_reactome)
}
