% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_CorrectBg_p.R
\name{Get_CorrectBg_p}
\alias{Get_CorrectBg_p}
\title{Get_CorrectBg_p}
\usage{
Get_CorrectBg_p(
  Single_data,
  scPagwas.TRS.Score,
  iters_singlecell,
  n_topgenes,
  scPagwas_topgenes,
  assay
)
}
\arguments{
\item{Single_data}{Seurat object}

\item{scPagwas.TRS.Score}{raw score}

\item{iters_singlecell}{number of iterations}

\item{n_topgenes}{number of genes}

\item{scPagwas_topgenes}{gene list}
}
\value{
correct_pdf
}
\description{
Get_CorrectBg_p
}
