% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_CorrectBg_p.R
\name{get_p_from_empi_null}
\alias{get_p_from_empi_null}
\title{get_p_from_empi_null}
\usage{
get_p_from_empi_null(v_t, v_t_null)
}
\arguments{
\item{v_t}{raw score}

\item{v_t_null}{control raw score}
}
\value{
p
}
\description{
get_p_from_empi_null
}
