% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Singlecell_heritability_contributions.R
\name{get_Pathway_sclm}
\alias{get_Pathway_sclm}
\title{get_Pathway_sclm}
\usage{
get_Pathway_sclm(
  pa_block,
  pca_scCell_mat,
  data_mat,
  rawPathway_list,
  snp_gene_df,
  n.cores = 1,
  backingpath,
  Rns
)
}
\arguments{
\item{pa_block}{single pathway blocks.}

\item{pca_scCell_mat}{Element in Pagwas.}

\item{data_mat}{Element in Pagwas.}

\item{rawPathway_list}{Element in Pagwas..}

\item{snp_gene_df}{Element in Pagwas.}

\item{n.cores}{cores for regression}

\item{backingpath}{file address for bk files, no "/" in the end.}

\item{Rns}{Random bk names}
}
\value{
lm result for pathway in single cell.
}
\description{
get_Pathway_sclm
}
