% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scPagwas-data.R
\docType{data}
\name{gtf_df}
\alias{gtf_df}
\title{gtf_df}
\format{
data.frame.
}
\source{
Generated from gencode.v34.annotation.gtf.gz
}
\description{
A gene annotation files from \code{
library("rtracklayer")
gtf_df<- rtracklayer::import("/share/pub/dengcy/GWAS_Multiomics/pagwas/data/gencode.v34.annotation.gtf.gz")
gtf_df <- as.data.frame(gtf)
gtf_df <- gtf_df[,c("seqnames","start","end","type","gene_name")]
gtf_df <- gtf_df[gtf_df$type=="gene",]
gtf_df<-gtf_df[,c(1,2,3,5)]
colnames(gtf_df)<-c("chrom", "start","end","label")
save(gtf_df,file = "/share/pub/dengcy/GWAS_Multiomics/pagwas/data/gtf_df.RData")
}
}
\examples{
data(gtf_df)
str(gtf_df)
}
