% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scPagwas-data.R
\docType{data}
\name{GWAS_summ_example}
\alias{GWAS_summ_example}
\title{GWAS summary data frame}
\format{
data.frame.
}
\source{
Generated from PBC_GWAS_UKBiobank_summary_final
}
\description{
A GWAS summary data from \code{
library(readr)
library(dplyr)
GWAS_liver<-read_delim("/share/pub/mayl/Sherlock/XiangbingYu_Primary_biliary_cholangitis/PBC_GWAS_UKBiobank_summary_final",delim="\t")
#select the specific coloumn names: chrom，pos，rsid，beta，se，maf
GWAS_liver<-GWAS_liver[,c("CHR","POS","SNP","p","SE")]
colnames(GWAS_liver)<-c("chrom","pos","rsid","p","se")
COVOD19<-read_delim("/share/pub/mayl/01_COVID19_GWAS_round4/COVID19_HGI_B2_ALL_leave_23andme_20201020.b37.txt",delim="\t")
COVOD19<-COVOD19[,c("all_meta_AF","rsid")]
colnames(COVOD19)<-c("maf","rsid")
##get beta
GWAS_liver$beta<-GWAS_liver$se*qchisq(GWAS_liver$p,1,lower.tail=F)
GWAS_liver_inner <- inner_join(GWAS_liver,COVOD19,by="rsid")
GWAS_liver_inner <- as.data.frame(GWAS_liver_inner)
GWAS_summ_example <- as.data.frame(GWAS_liver_inner[sample(nrow(GWAS_liver_inner),10000),])
save(GWAS_summ_example,file="/share/pub/dengcy/GWAS_Multiomics/pagwas/data/GWAS_summ_example.RData")
}
}
\examples{
data(GWAS_summ_example)
str(GWAS_summ_example)
}
