% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scPagwas-data.R
\docType{data}
\name{ld_folder}
\alias{ld_folder}
\title{ld_folder
use vcftools to get ped and map files
LD data from \code{
/share/apps/vcftools/bin/vcftools --vcf /share/pub/dengcy/Singlecell/COVID19/1000genomes_all_genotypes.vcf --plink-tped --out /share/pub/dengcy/Singlecell/COVID19/1000genomes_all_genotypes}
\format{
vector
}
\source{
Generated from PLINK 1.90 linux
}
\usage{
ld_folder
}
\description{

}
\examples{
data(ld_folder)
str(ld_folder)
}
\keyword{datasets}
