% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Link_pathway_blocks_gwas.R
\name{Link_pathway_blocks_gwas}
\alias{Link_pathway_blocks_gwas}
\title{Link_pathway_blocks_gwas}
\usage{
Link_pathway_blocks_gwas(Pagwas, chrom_ld = NULL, singlecell = T, celltype = T)
}
\arguments{
\item{Pagwas}{Pagwas format, deault is NULL.}

\item{chrom_ld}{LD data for 22 chromosome.}

\item{singlecell}{whether to run singlecell progress}

\item{celltype}{whether to run celltype progress}
}
\value{
Pagwas list of blocks for snp to gene and pathway score.
}
\description{
Link pathway blocks region and gwas summary snps
Takes block information, potentially independent LD pathway_blocks
or gene blocks,
}
\author{
Chunyu Deng
}
\keyword{Link_pathway_blocks_gwas,}
\keyword{calculate}
\keyword{data.}
\keyword{for}
\keyword{matrix}
\keyword{singlecell}
\keyword{svd}
\keyword{the}
