% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Link_pathway_blocks_gwas.R
\name{link_pwpca_block}
\alias{link_pwpca_block}
\title{link_pwpca_block}
\usage{
link_pwpca_block(
  pa_block,
  pca_cell_df,
  merge_scexpr,
  snp_gene_df,
  rawPathway_list
)
}
\arguments{
\item{pa_block}{snp blocks for pathway.}

\item{pca_cell_df}{pca matrix for cell types.}

\item{merge_scexpr}{mean expression for merged cell types of single cell
data.}

\item{snp_gene_df}{Data frame of snp mapping to gene.}

\item{rawPathway_list}{Raw pathway list before filter}
}
\value{
pathway block including "x" "y" "snps" "include_in_inference"
}
\description{
Link the pca score and expression for each pathway genes
for each block
Requires rownames that are identitcal to block labels loaded previously.
}
