% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Link_pathway_blocks_gwas.R
\name{make_ld_matrix}
\alias{make_ld_matrix}
\title{make_ld_matrix}
\usage{
make_ld_matrix(all_snps, ld_data)
}
\arguments{
\item{all_snps}{The snps that were queried}

\item{ld_data}{A returned LD matrix with SNP}
}
\description{
construct the final matrix for regression in blocks
}
