% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_functions.R
\name{mean_expr}
\alias{mean_expr}
\title{mean_expr}
\usage{
mean_expr(Pagwas)
}
\arguments{
\item{Pagwas}{Pagwas data list}
}
\value{

}
\description{
To get the mean expression for single cell data.
}
\examples{
library(scPagwas)
Pagwas$merge_scexpr <- mean_expr(Pagwas)

}
