% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_CorrectBg_p.R
\name{Merge_celltype_p}
\alias{Merge_celltype_p}
\title{Merge_celltype_p}
\usage{
Merge_celltype_p(single_p, celltype)
}
\arguments{
\item{single_p}{single p}

\item{celltype}{celltype}
}
\value{
celltype_p
}
\description{
Merge_celltype_p
}
