% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scPagwas_main.R
\name{merge_pagwas}
\alias{merge_pagwas}
\title{merge_pagwas}
\usage{
merge_pagwas(Pagwas_list = NULL, n_topgenes = 1000)
}
\arguments{
\item{Pagwas_list}{list format, element including seruat format result.}

\item{n_topgenes}{(integr)Number of top associated gene selected to calculate the scPagwas score;}
}
\value{

}
\description{
Merging several pagwas result to one result.When your single
cell data is too large to run, you can split your data into several
split, then use this function to merge them. Split the data randomly or
in order of cellytpes.
}
\note{
1.The rownames genes and gwas data must uniform.
2.The merge_pagwas function only can use to merge the seruat format result.
That is "singlecell = T;celltype =F" can use to all the split data,
"singlecell = T;celltype = T" can be used only when the data is splited by celltypes.
but "singlecell = F;celltype =T" can not be used.
}
\examples{

Pagwas_merge <- merge_pagwas(Pagwas_list = list(Pagwas_data1, Pagwas_data2))
}
