% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Celltype_heritability_contributions.R
\name{Pa_Pagwas_perform_regression}
\alias{Pa_Pagwas_perform_regression}
\title{Pa_Pagwas_perform_regression}
\usage{
Pa_Pagwas_perform_regression(pa_block)
}
\arguments{
\item{pa_block}{Pagwas format, deault is NULL.}
}
\value{

}
\description{
Run regression
}
\examples{
library(scPagwas)
Pagwas <- Pagwas_perform_regression(Pagwas, iters = 200)
}
