% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Celltype_heritability_contributions.R
\name{pa_parameter_regression}
\alias{pa_parameter_regression}
\title{pa_parameter_regression}
\usage{
pa_parameter_regression(vectorized_Pagwas_data)
}
\arguments{
\item{vectorized_Pagwas_data}{Pagwas data that has been vectorized}
}
\value{

}
\description{
Find parameter estimates for the data.
}
