% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Link_pathway_blocks_gwas.R
\name{Pachrom_func}
\alias{Pachrom_func}
\title{Title}
\usage{
Pachrom_func(Pagwas, Pachrom_block_list, chrom_gwas_list)
}
\arguments{
\item{chrom_gwas_list}{}
}
\value{

}
\description{
Title
}
