% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pagwas_main.R
\name{Pagwas_main}
\alias{Pagwas_main}
\title{Main wrapper functions for celltypes.}
\usage{
Pagwas_main(
  Pagwas = NULL,
  gwas_data = NULL,
  add_eqtls = "OnlyTSS",
  eqtls_files = NULL,
  eqtls_cols = c("rs_id_dbSNP151_GRCh38p7", "variant_pos", "tss_distance", "gene_chr",
    "gene_start", "gene_end", "gene_name", "pval_beta"),
  block_annotation = NULL,
  Single_data = NULL,
  FilterSingleCell = FALSE,
  nfeatures = NULL,
  Pathway_list = NULL,
  chrom_ld = NULL,
  marg = 10000,
  maf_filter = 0.01,
  min.reads = 5,
  min.detected = 1,
  min.lib.size = 200,
  min_clustercells = 10,
  min.pathway.size = 5,
  max.pathway.size = 300,
  iters = 200,
  perform_cv = F,
  n_folds = 10,
  n.cores = 1,
  simp_results = F,
  ...
)
}
\arguments{
\item{Pagwas}{(list)Pagwas data list, default is "NULL"}

\item{gwas_data}{(data.frame)GWAS summary data}

\item{add_eqtls}{(character)There are three options: "OnlyTSS","OnlyEqtls","Both"; "OnlyTSS" means there is only snps within TSS window;"OnlyEqtls" represents only significant snps in eqtls files;  "Both" represents both significant snps in eqtls and the other snps in GWAS summary files;}

\item{eqtls_files}{(character)eqtls files names,frome GTEx.}

\item{eqtls_cols}{(character) the needed columns in eqtls_files; In our examples files, the columns including: c("rs_id_dbSNP151_GRCh38p7","variant_pos","tss_distance","gene_chr", "gene_start", "gene_end","gene_name")}

\item{block_annotation}{(data.frame) Start and end points for block traits, usually genes.}

\item{Single_data}{(Seruat)Input the Single data in seruat format, Idents should be the celltypes annotation.}

\item{FilterSingleCell}{(logical)whether to filter the single cell data.if you
filter it before,choose FALSE, otherwise set TRUE.}

\item{nfeatures}{(integr) The parameter for FindVariableFeatures, NULL means select all genes}

\item{Pathway_list}{(list,character) pathway gene sets list}

\item{chrom_ld}{(list,numeric)LD data for 22 chromosome.}

\item{marg}{(integr)gene-TSS-window size}

\item{maf_filter}{(numeric)Filter the maf, default is 0.01}

\item{min.reads}{(integr)Only use is when FilterSingleCell is TRUE.Threshold for total reads fo each cells. default is 5}

\item{min.detected}{(integr)Only use is when FilterSingleCell is TRUE.Threshold for total cells fo each gene. default is 1}

\item{min.lib.size}{(integr)Only use is when FilterSingleCell is TRUE.Threshold for single data library. default is 200}

\item{min_clustercells}{(integr)Only use is when FilterSingleCell is TRUE.Threshold for total cells fo each cluster.default is 10}

\item{min.pathway.size}{(integr)Threshold for min pathway gene size. default is 5}

\item{max.pathway.size}{(integr)Threshold for max pathway gene size. default is 300}

\item{iters}{(integr)number of bootstrap iterations to perform, default is 200}

\item{perform_cv}{(logical)whether to Perform regularization inference.}

\item{n_folds}{(integr)folds for regularized inference,default is 10.}

\item{n.cores}{(integr)Parallel cores,default is 1. use detectCores() to check the cores in computer.}

\item{simp_results}{(logical)Whether to return a simple results.If TURE,there is only finaly result return; if FALSE, all the Intermediate Data will keep in pagwas, this option is used to save time when rerun the function.}

\item{...}{other parameters for subfunction}
}
\value{

}
\description{
Main Pagwas wrapper function in order.
}
\details{
The entry point for Pagwas analysis.
}
\examples{
library(scPagwas)
data(Genes_by_pathway_kegg)
data(GWAS_summ_example)
data(gtf_df)
data(scRNAexample)
#data(eqtls_files)
data(chrom_ld)
#1. OnlyTSS
Pagwas<-Pagwas_main(Pagwas = NULL,
                    gwas_data = GWAS_summ_example,
                    add_eqtls="OnlyTSS",
                    block_annotation = gtf_df,
                    Single_data = scRNAexample,
                    Pathway_list=Genes_by_pathway_kegg,
                    chrom_ld = chrom_ld)
#1. Both
scPagwas_main <- function(Pagwas = NULL,
                          gwas_data = GWAS_summ_example,
                          add_eqtls="Both",
                          eqtls_files=eqtls_files,
                          eqtl_p=0.05,
                          block_annotation = gtf_df,
                          Single_data = scRNAexample,
                          Pathway_list=Genes_by_pathway_kegg,
                          chrom_ld = chrom_ld)

}
