% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Celltype_heritability_contributions.R
\name{Pagwas_perform_regularized_regression}
\alias{Pagwas_perform_regularized_regression}
\title{Pagwas_perform_regularized_inference}
\usage{
Pagwas_perform_regularized_regression(Pagwas, n_folds = 10)
}
\arguments{
\item{Pagwas}{Pagwas data}

\item{n_folds}{folds for regularized inference}
}
\value{

}
\description{
Run inference with added regularization.
If p-values are desired use the other inference function. This for
prediction purposes.
}
