% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pagwas_perform_inference.R
\name{para_names_adjust}
\alias{para_names_adjust}
\title{para_names_adjust}
\usage{
para_names_adjust(Pagwas, lm_results = Pagwas$lm_results)
}
\arguments{
\item{Pagwas}{Pagwas format}

\item{lm_results}{Pagwas data list for lm_results, it is the result of lm}
}
\value{

}
\description{
Adjustment the names of parameters and pca_cell_df, for the unidentified signals;
}
