% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pagwas_perform_inference.R
\name{Parameter_regression}
\alias{Parameter_regression}
\title{Parameter_regression}
\usage{
Parameter_regression(vectorized_Pagwas_data)
}
\arguments{
\item{vectorized_Pagwas_data}{Pagwas data that has been vectorized}
}
\value{

}
\description{
Find parameter estimates for the data.
}
