% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pathway_annotation_input.R
\name{Pathway_annotation_input}
\alias{Pathway_annotation_input}
\title{Pathway_annotation_input}
\usage{
Pathway_annotation_input(Pagwas, block_annotation)
}
\arguments{
\item{Pagwas}{Pagwas format data list inherit from other functions.}

\item{block_annotation}{Start and end points for block traits,
usually genes.}
}
\value{
list for pathway blocks
}
\description{
Deal with the pathway genes with snp blocks
Initial the pathway blocks for mapping the pathway to snp.
}
\author{
Chunyu Deng
}
\keyword{Initial}
\keyword{Pathway_annotation_input,}
\keyword{blocks}
\keyword{for}
\keyword{mapping}
\keyword{pathway}
\keyword{snp.}
\keyword{the}
\keyword{to}
