% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pathway_pcascore_run.R
\name{Pathway_pcascore_run}
\alias{Pathway_pcascore_run}
\title{Pathway_pcascore_run}
\usage{
Pathway_pcascore_run(
  Pagwas = NULL,
  Pathway_list = NULL,
  min.pathway.size = 10,
  max.pathway.size = 1000
)
}
\arguments{
\item{Pagwas}{Pagwas format}

\item{Pathway_list}{Pathawy gene list(gene symbol),list name is pathway name.}

\item{min.pathway.size}{Threshold for min pathway gene size.}

\item{max.pathway.size}{Threshold for max pathway gene size.}
}
\value{

}
\description{
The function can calculate the pathway pcascore for celltypes and single cells.
}
\examples{
library(scPagwas)
Pagwas <- list()
#Start to read the single cell data
Single_data <- readRDS(system.file("extdata", "scRNAexample.rds", package = "scPagwas"))
Pagwas <- Single_data_input(
  Pagwas = Pagwas,
  assay = "RNA",
  Single_data = Single_data,
  Pathway_list = Genes_by_pathway_kegg
)
Single_data <- Single_data[, colnames(Pagwas$data_mat)]
#Run pathway pca score!
Pagwas <- Pathway_pcascore_run(
  Pagwas = Pagwas,
  Pathway_list = Genes_by_pathway_kegg
)
}
\author{
Chunyu Deng
}
\keyword{Single_data_input,}
\keyword{calculate}
\keyword{data.}
\keyword{for}
\keyword{matrix}
\keyword{singlecell}
\keyword{svd}
\keyword{the}
