% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Singlecell_heritability_contributions2.0.R
\name{Pathway_singlecell_rankp}
\alias{Pathway_singlecell_rankp}
\title{link_scCell_pwpca_block}
\usage{
Pathway_singlecell_rankp(Pagwas, n.cores = 1, remove_outlier = T)
}
\arguments{
\item{Pagwas}{Pagwas format, deault is NULL.}

\item{n.cores}{Parallel cores,default is 1. use detectCores() to
check the cores in computer.}

\item{remove_outlier}{Whether to remove the outlier for scPagwas score.}
}
\value{

}
\description{
link the single cell pca score and expression for each
pathway genes for each block(single cell)
construct the final matrix for regression in blocks
}
\examples{
library(scPagwas)
}
