% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pathway_pcascore_run.R
\name{PathwayPCAtest}
\alias{PathwayPCAtest}
\title{PathwayPCAtest}
\usage{
PathwayPCAtest(Pathway_list, scCounts, n.cores = 1)
}
\arguments{
\item{Pathway_list}{Pathawy gene list(gene symbol),list name is pathway name.}

\item{scCounts}{single cell Counts pathway}

\item{n.cores}{Parallel cores,default is 1. use detectCores() to check the cores in computer.}
}
\value{

}
\description{
Calculate pca score
}
