% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bar_positie_nagtive.R
\name{plot_bar_positie_nagtive}
\alias{plot_bar_positie_nagtive}
\title{plot_bar_positie_nagtive}
\usage{
plot_bar_positie_nagtive(
  seurat_obj,
  var_ident,
  var_group,
  vec_group_colors = NULL,
  f_color = grDevices::colorRampPalette(RColorBrewer::brewer.pal(n = 11, name =
    "RdYlBu")),
  do_plot = FALSE,
  title = NULL,
  p_thre = 0.05,
  fontsize_title = 24,
  fontsize_axistitle_x = 18,
  fontsize_axistitle_y = 18,
  fontsize_axistext_x = 12,
  fontsize_axistext_y = 12,
  fontsize_legendtitle = 12,
  fontsize_legendtext = 10,
  aspect.ratio = 1.2,
  output.prefix = "Test",
  output.dirs = NULL,
  width = 7,
  height = 7
)
}
\arguments{
\item{seurat_obj}{Seurat object (Seurat ^3.0).}

\item{var_ident}{the identify variable, character.}

\item{var_group}{the group variable, character.}

\item{vec_group_colors}{a vector of colors, named by corresponding group.
Length must match number of groups. Character.}

\item{f_color}{if vec_group_colors is not provided, the user may instead provide a
function f_color() that takes as its only argument the number of colors to generate.}

\item{do_plot}{Whether to plot, logical.}

\item{title}{NULL to leave out.}

\item{p_thre}{threshold for p value.}

\item{fontsize_title}{NULL to leave out.}

\item{fontsize_axistitle_x}{NULL to leave out.}

\item{fontsize_axistitle_y}{NULL to leave out.}

\item{fontsize_axistext_x}{NULL to leave out.}

\item{fontsize_axistext_y}{NULL to leave out.}

\item{fontsize_legendtitle}{NULL to leave out.}

\item{fontsize_legendtext}{NULL to leave out.}

\item{aspect.ratio}{default is 1.2.}

\item{output.prefix}{prefix for output files.}

\item{output.dirs}{directory for files.}

\item{width}{figure width.}

\item{height}{figure height.}
}
\value{
a ggplot2 object
}
\description{
Generate barplot of identity group composition
Thanks to the SCOPfunctions package
https://github.com/CBMR-Single-Cell-Omics-Platform/SCOPfunctions/blob/main/R/plot.R
Generate a percentage barplot that shows the composition of each identity
(e.g. sample)in terms of groups (e.g. positive and negative cells for scPagwas)
}
\examples{
load(system.file("extdata", "Pagwas_data.RData", package = "scPagwas"))
plot_bar_positie_nagtive(
  seurat_obj = Pagwas_data,
  var_ident = "anno",
  var_group = "positiveCells",
  vec_group_colors = c("#E8D0B3", "#7EB5A6"),
  do_plot = FALSE
)
}
\keyword{bars}
\keyword{cellytpes.}
\keyword{for}
\keyword{plot}
\keyword{plot_bar_positie_nagtive,}
\keyword{pvalue}
\keyword{the}
