% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pathway_contribution_network.R
\name{plot_pathway_contribution_network}
\alias{plot_pathway_contribution_network}
\title{plot_pathway_contribution_network}
\usage{
plot_pathway_contribution_network(
  mat_datExpr,
  vec_pathwaycontribution,
  Celltype_anno,
  pathways_thre = 0.05,
  adjacency.power = 4,
  adjacency.thre = 0.1,
  n_max_pathways = 10,
  igraph_algorithm = "drl",
  fontface_labels = "bold.italic",
  color_point = pal_d3(alpha = 0.5)(10),
  color_edge = "#9D9D9D",
  fontSize_label_lg = 3,
  fontSize_legend_lg = 1,
  fontSize_legend_xlg = 1,
  limits = c(0, 10),
  edge_thickness = 2,
  do_plot = F,
  figurenames = NULL,
  width = 7,
  height = 7
)
}
\arguments{
\item{mat_datExpr}{data matrix for single cell data}

\item{vec_pathwaycontribution}{vector of pathway contribution score from pagwas}

\item{n_max_pathways}{number of max pathways for showing.}

\item{igraph_algorithm}{defalt is drl,method for igragh,details see igraph packages.}

\item{fontface_labels}{fontface for labels}

\item{color_edge}{edge color.}

\item{fontSize_label_lg}{fontSize for lg label}

\item{fontSize_legend_lg}{fontSize for lg legend}

\item{fontSize_legend_xlg}{fontSize for xlg legend}

\item{edge_thickness}{thickness of edge}

\item{do_plot}{Whether to plot, logical}

\item{figurenames}{The filename and address of the output plot}

\item{width}{figure width}

\item{height}{figure height}

\item{vec_pathways_highlight}{vector of pathway highlight from your choose.}
}
\description{
plot the network based on pathway contribution
oaqc

Thanks to the SCOPfunctions package
https://github.com/CBMR-Single-Cell-Omics-Platform/SCOPfunctions/blob/main/R/plot.R
}
\examples{
plot_pathway_contribution_network(
mat_datExpr=pca_cell_df,
vec_pathwaycontribution=Pagwas$Pathway_block_heritability,
vec_pathways_highlight=names(sort(Pagwas$Pathway_block_heritability,decreasing = T)[1:5]),
n_max_pathways=20,
igraph_algorithm = "drl",
fontface_labels="bold.italic",
color_edge = "#9D9D9D",
fontSize_label_lg=4,
fontSize_legend_lg=4,
fontSize_legend_xlg=4,
edge_thickness = 1
)
}
