% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scpathway_contri_dot.R
\name{plot_scpathway_dot}
\alias{plot_scpathway_dot}
\title{plot_scpathway_dot}
\usage{
plot_scpathway_dot(
  Pagwas,
  topn_path_celltype = 10,
  filter_p = 0.05,
  max_logp = 10,
  display_max_sizes = F,
  size_var = "lm_beta",
  col_var = "logrankPvalue",
  shape.scale = 8,
  cols.use = c("lightgrey", "#E45826"),
  dend_x_var = "lm_beta",
  dist_method = "euclidean",
  hclust_method = "ward.D",
  do_plot = F,
  figurenames = NULL,
  width = 7,
  height = 7,
  ...
)
}
\arguments{
\item{topn_path_celltype}{number specific pahtways to select for each celltypes.}

\item{filter_p}{0.05 threshold for pvalues}

\item{display_max_sizes}{Boolean : Display max shape size behind each shape ? (Default=TRUE)}

\item{size_var}{#'   If numeric : Column/List index which control shape sizes. This column/element has to be numeric.
#'   Can also be a column/element name or a vector of the same size than the input dataset.
#'   Set to NA if you don't want to control shape size.}

\item{col_var}{#'   If numeric : Column/List index which control shape colors.
#'   Can also be a column/element name or a vector of the same size than the input dataset.
#'   Set to NA if you don't want to control shape color.}

\item{shape.scale}{Scale the size of the shapes, similar to cex.}

\item{cols.use}{1 color or a vector containing multiple colors to color shapes.}

\item{dend_x_var}{A vector containing Column/List indexes or Column/List names to compute the x axis dendrogramm.}

\item{dist_method}{The distance measure to be used. This must be one of "euclidean", "maximum", "manhattan", "canberra", "binary" or "minkowski".}

\item{hclust_method}{The agglomeration method to be used. This must be one of "single", "complete", "average", "mcquitty", "ward.D", "ward.D2", "centroid" or "median".}

\item{do_plot}{Boolean : whether to print the plot}

\item{...}{other parameters for plot_scpathway_contri_dot}
}
\value{

}
\description{
plot_scpathway_dot
}
