% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scPagwas_Visualization.R
\name{preprogress_PaSingle_data}
\alias{preprogress_PaSingle_data}
\title{preprogress_PaSingle_data}
\usage{
preprogress_PaSingle_data(Single_data, npcs = 10, nfeatures = 20)
}
\arguments{
\item{nfeatures}{}
}
\value{

}
\description{
preprogress_PaSingle_data
}
