% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_functions.R
\name{qvalue}
\alias{qvalue}
\title{Estimate the q-values for a given set of p-values}
\usage{
qvalue(p, fdr.level = NULL, pfdr = FALSE, lfdr.out = TRUE, pi0 = NULL, ...)
}
\arguments{
\item{p}{A vector of p-values (only necessary input).}

\item{fdr.level}{A level at which to control the FDR. Must be in (0,1]. Optional; if this is
selected, a vector of TRUE and FALSE is returned that specifies
whether each q-value is less than fdr.level or not.}

\item{pfdr}{An indicator of whether it is desired to make the
estimate more robust for small p-values and a direct finite sample estimate of pFDR -- optional.}

\item{lfdr.out}{If TRUE then local false discovery rates are returned. Default is TRUE.}

\item{pi0}{It is recommended to not input an estimate of pi0. Experienced users can use their own methodology to estimate
the proportion of true nulls or set it equal to 1 for the BH procedure.}

\item{\ldots}{Additional arguments passed to \code{\link{pi0est}} and \code{\link{lfdr}}.}
}
\value{

}
\description{
Estimate the q-values for a given set of p-values.  The q-value of a
test measures the proportion of false positives incurred (called the
false discovery rate) when that particular test is called significant.
}
\details{
The function \code{\link{pi0est}} is called internally and calculates the estimate of \eqn{\pi_0}{pi_0},
the proportion of true null hypotheses. The function \code{\link{lfdr}} is also called internally and
calculates the estimated local FDR values.  Arguments for these functions can be included via \code{...} and
will be utilized in the internal calls made in \code{\link{qvalue}}. See \url{http://genomine.org/papers/Storey_FDR_2011.pdf}
for a brief introduction to FDRs and q-values.
}
\examples{
# import data
data(hedenfalk)
p <- hedenfalk$p

# get q-value object
qobj <- qvalue(p)
plot(qobj)
hist(qobj)

# options available
qobj <- qvalue(p, lambda=0.5, pfdr=TRUE)
qobj <- qvalue(p, fdr.level=0.05, pi0.method="bootstrap", adj=1.2)

}
\references{
Storey JD. (2011) False discovery rates. In \emph{International Encyclopedia of Statistical Science}. \cr
\url{http://genomine.org/papers/Storey_FDR_2011.pdf} \cr
\url{http://www.springer.com/statistics/book/978-3-642-04897-5}
}
\author{
John D. Storey
}
\keyword{qvalue}
