% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_functions.R
\name{qWishartMax}
\alias{qWishartMax}
\title{qWishartMax
the source code from RMTstat package}
\usage{
qWishartMax(p, ndf, pdim, var = 1, beta = 1, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{log.p}{}
}
\value{

}
\description{
qWishartMax
the source code from RMTstat package
}
