% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Singlecell_heritability_contributions.R
\name{Random_PCC}
\alias{Random_PCC}
\title{Random_PCC}
\usage{
Random_PCC(gPas, datamat, seed = 1234, random_times = 100, select_num = 10000)
}
\arguments{
\item{gPas}{gPas score}

\item{datamat}{data matrix for single cell}

\item{seed}{random seed}

\item{random_times}{random times}

\item{select_num}{the cells for each random}
}
\description{
Random_PCC
}
