% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_functions.R
\name{rankPvalue}
\alias{rankPvalue}
\title{rankPvalue}
\usage{
rankPvalue(
  datS,
  columnweights = NULL,
  na.last = "keep",
  ties.method = "average",
  calculateQvalue = TRUE,
  pValueMethod = "all"
)
}
\arguments{
\item{datS}{matrix or data.frame}

\item{columnweights}{NULL}

\item{na.last}{"keep"}

\item{ties.method}{"average"}

\item{calculateQvalue}{TRUE}

\item{pValueMethod}{"all"}
}
\value{

}
\description{
rankPvalue
}
