% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Singlecell_heritability_contributions.R
\name{scBoot_evaluate}
\alias{scBoot_evaluate}
\title{scBoot_evaluate}
\usage{
scBoot_evaluate(Pagwas, bootstrap_iters = bootstrap_iters, n.cores = n.cores)
}
\arguments{
\item{Pagwas}{Pagwas object}

\item{bootstrap_iters}{number of bootstrap iterations to run}

\item{n.cores}{cores}
}
\value{

}
\description{
Bootstrap to evaluate for confidence intervals.
}
