% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Single_data_input.R
\name{scCount_data_input}
\alias{scCount_data_input}
\title{scCount_data_input}
\usage{
scCount_data_input(
  Pagwas,
  count_data,
  meta_data,
  Pathway_list,
  min_clustercells
)
}
\arguments{
\item{Pagwas}{Pagwas format}

\item{count_data}{(data.frame or matrix) count data of your single cell data to use, the rownames is gene names, the colnames is cell names.}

\item{meta_data}{(data.frame) meta data of your single cell data to use, the first column is cell names. the cell names should be the same as the first row of count_data. It should include the "cell_type" column.}

\item{Pathway_list}{(list,character) pathway gene sets list}

\item{min_clustercells}{Threshold for total cells fo each cluster.}
}
\value{
Pagwas list including:
"Celltype_anno"    "data_mat"  "merge_scexpr"
}
\description{
Input the scCount data in dataframe or matrix, the rownames is gene names, the colnames is cell names.
}
\examples{
library(scPagwas)

}
\author{
Chunyu Deng
}
\keyword{cell}
\keyword{extract}
\keyword{matrix.}
\keyword{scCount_data_input,}
\keyword{single}
\keyword{the}
