% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scGene_rankP.R
\name{scGene_rankP}
\alias{scGene_rankP}
\title{rankPvalue}
\usage{
scGene_rankP(Single_mat, na.last = "keep", ties.method = "average")
}
\arguments{
\item{Single_mat}{matrix or data.frame}

\item{na.last}{"keep"}

\item{ties.method}{"average"}
}
\value{
rank p value for expression matrix.
}
\description{
rankPvalue
}
