% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Singlecell_heritability_contributions.R
\name{scGet_bootresults_df}
\alias{scGet_bootresults_df}
\title{scGet_bootresults_df}
\usage{
scGet_bootresults_df(value_collection, annotations, model_estimates)
}
\arguments{
\item{value_collection}{collection of bootstrapped value estimates}

\item{annotations}{vector of annotation names}

\item{model_estimates}{estimates for bias parameter estimates}
}
\value{

}
\description{
Helper function to make a summary table of results from bootstrap data.
}
