% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Singlecell_heritability_contributions.R
\name{scGet_GEI}
\alias{scGet_GEI}
\title{scGet_GEI}
\usage{
scGet_GEI(scPagwas.gPAS.score, data_mat)
}
\arguments{
\item{scPagwas.gPAS.score}{score of scPagwas pathway lm}

\item{data_mat}{matrix for single cell data}
}
\value{
result list including gene heritability correlation
}
\description{
scGet_GEI
}
