% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scPagwas_perform_score.R
\name{scGet_gene_heritability_contributions}
\alias{scGet_gene_heritability_contributions}
\title{scGet_gene_heritability_contributions}
\usage{
scGet_gene_heritability_contributions(scPagwas_score, data_mat)
}
\arguments{
\item{data_mat}{}
}
\value{

}
\description{
scGet_gene_heritability_contributions
}
