% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Singlecell_heritability_contributions2.0.R
\name{scGet_Pathway_heritability_contributions}
\alias{scGet_Pathway_heritability_contributions}
\title{scGet_Pathway_heritability_contributions}
\usage{
scGet_Pathway_heritability_contributions(pca_scCell_mat, Pathway_sclm_results)
}
\arguments{
\item{pca_scCell_mat}{pca score matrix}

\item{parameters}{parameter fit}
}
\value{

}
\description{
predicting the importance of a pathway based on its pcascore
}
