% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Singlecell_heritability_contributions.R
\name{scGet_PCC}
\alias{scGet_PCC}
\title{scGet_PCC}
\usage{
scGet_PCC(scPagwas.gPAS.score, data_mat)
}
\arguments{
\item{scPagwas.gPAS.score}{score of scPagwas pathway lm}

\item{data_mat}{matrix for single cell data}
}
\value{
result list including pearson correlation coefficients(PCC)
}
\description{
scGet_PCC
}
