% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Singlecell_heritability_contributions.R
\name{scGet_PCC2}
\alias{scGet_PCC2}
\title{scGet_PCC2}
\usage{
scGet_PCC2(Pagwas, random_times = 100)
}
\arguments{
\item{Pagwas}{a result list including scPagwas.gPAS.score and data_mat}

\item{random_times}{the random times}
}
\description{
scGet_PCC2
}
