% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scPagwas_main.R
\name{scPagwas_main}
\alias{scPagwas_main}
\title{Main wrapper functions for scPagwas}
\usage{
scPagwas_main(
  Pagwas = NULL,
  gwas_data = NULL,
  output.prefix = "Test",
  output.dirs = "scPagwastest_output",
  block_annotation = NULL,
  Single_data = NULL,
  assay = c("RNA", "SCT"),
  Pathway_list = NULL,
  chrom_ld = NULL,
  split_n = 1,
  marg = 10000,
  maf_filter = 0.01,
  min_clustercells = 10,
  min.pathway.size = 5,
  max.pathway.size = 300,
  iters = 200,
  n_topgenes = 1000,
  singlecell = T,
  celltype = T,
  seruat_return = T,
  remove_outlier = T,
  ncores = 1
)
}
\arguments{
\item{Pagwas}{(list)default is "NULL" when you first run the funciton; Pagwas
should be list class; Sometimes, It can inherit the result from "scPagwas_main" function
last time, when you turn the "seruat_return" to FALSE; It is suitable for circulation
running for the same single data.}

\item{gwas_data}{(data.frame)GWAS summary data; It must have some colmuns such as:
chrom|    pos    |   rsid    |   se  |  beta |  maf
6 | 119968580 | rs1159767 | 0.032 | 0.019 |0.5275
10 | 130566523 |  rs559109 | 0.033 | 0.045 |0.4047
5 | 133328825 | rs6893145 | 0.048 | 0.144 |0.1222
7 | 146652932 | rs13228798| 0.035 | 0.003 | 0.3211}

\item{output.prefix}{(character)output.prefix,prefix used for output tables;}

\item{output.dirs}{(character)output directory for result; If the directory is nonexistence,
the function will create it;}

\item{block_annotation}{(data.frame) Start and end position for block traits, usually genes.}

\item{Single_data}{(character or Seruat)Input the Single data in seruat format,
or the Seruat data address for rds format;the "Idents" should be the celltypes
annotation, if the "celltypes" is TRUE.}

\item{assay}{(character)assay data of your single cell data to use,"RNA" or "SCT",default is "RNA";}

\item{Pathway_list}{(list,character) pathway gene sets list,the name of list is pahtway names;
the element of list including genes;}

\item{chrom_ld}{(list,numeric)LD data for 22 chromosomes. Processed data is \code{data(chrom_ld)}}

\item{split_n}{(integr)Number of blocks which are split from singlecell data;
sometimes,the data is too big to run; we also advice split the data before run the main function.}

\item{marg}{(integr) the distance to TSS site,default is 10000, then gene-TSS-window size is 20000.}

\item{maf_filter}{(numeric)Filter the "maf" for gwas_data, default is 0.01;}

\item{min_clustercells}{(integr)Threshold for number of cells in each cluster.default is 10;}

\item{min.pathway.size}{(integr)Threshold for min pathway gene size. default is 5}

\item{max.pathway.size}{(integr)Threshold for max pathway gene size. default is 300}

\item{iters}{(integr)Number of bootstrap iterations to perform, default is 200;}

\item{n_topgenes}{(integr)Number of top associated gene selected to calculate the scPagwas score;}

\item{singlecell}{(logical)Whether to produce the singlecell result;}

\item{celltype}{(logical)Whether to produce the celltypes result;}

\item{seruat_return}{(logical) Whether return the Seruat format result, if not,will return a list result;}

\item{remove_outlier}{(logical)Whether to remove the outlier for scPagwas score.}

\item{ncores}{(integr)Parallel cores,default is 1. use \code{detectCores()} to check the cores in computer.}
}
\value{
Returns a Seruat data with entries(seruat_return=T):
\describe{
\position{assay:}{
\item{scPagwasPaPca:}{An assay for S4 type of data; the svd result for pathways in each cells;}}

\position{meta.data}{
\item{scPagwas.lmtopgenes.Score1:}{ the column for "meta.data";Enrichment socre for inheritance associated top genes.}
\item{sclm_score:}{ the column for "meta.data";Inheritance regression effects for each cells}}

\position{misc: element in result,\code{Pagwas@misc }}{
\item{Pathway_list:}{a list for pathway gene list intersecting with single cell data}
\item{pca_cell_df:}{ a data frame for pathway pca result for each celltype.}
\item{sclm_results:}{ the regression result for each cell.}
\item{allsnp_gene_heritability_correlation:}{
heritability correlation for each gene;}
\item{Pathway_ctlm_results:}{ the regression result for each pathway in each celltype}
\item{lm_results:}{ the regression result for celltypes}
\item{Pathway_ct_results:}{Inheritance conmtribution Matrix for pathways and celltypes, it contribute from "Pathway_ctlm_results"}
}

}
Returns files:

\describe{
\item{d:}{ max(nu, nv) approximate singular values}
\item{u:}{ nu approximate left singular vectors (only when right_only=FALSE)}
\item{v:}{ nv approximate right singular vectors}
\item{iter:}{ The number of Lanczos iterations carried out}
\item{mprod:}{ The total number of matrix vector products carried out}
}

Returns a Seruat data with entries(seruat_return=T):
\describe{
\item{scPagwasPaPca:}{Assays for S4 type of data; the svd result for pathways in each cells;}
\item{scPagwas.lmtopgenes.Score1:}{ the column for "meta.data";Enrichment socre for inheritance associated top genes.}
\item{sclm_score:}{ the column for "meta.data";Inheritance regression effects for each cells}
\item{Pathway_list:}{ The number of Lanczos iterations carried out}
\item{pca_cell_df:}{ The total number of matrix vector products carried out}
\item{sclm_results:}{ The total number of matrix vector products carried out}
\item{allsnp_gene_heritability_correlation:}{ The total number of matrix vector products carried out}
\item{Pathway_ctlm_results:}{ The total number of matrix vector products carried out}
\item{lm_results:}{ The total number of matrix vector products carried out}
\item{Pathway_ct_results:}{ The total number of matrix vector products carried out}
}
}
\description{
Main Pagwas wrapper for workflow function.
}
\details{
scPagwas, A single-cell pathway-based principal component (PC)-scoring algorithm named
for integrating polygenic signals from GWAS with single cell data
to infer disease-relevant cell populations and score the associations of
individual cells with complex diseases.

The entry point for Pagwas analysis.
}
\note{
1.When you run the package in linux server, you can run
\code{export OPENBLAS_NUM_THREADS=1}
before enter into R environment;
}
\examples{
library(scPagwas)
data(Genes_by_pathway_kegg)
data(GWAS_summ_example)
data(gtf_df)
data(scRNAexample)
# data(eqtls_files)
data(chrom_ld)
# 1. OnlyTSS
Pagwas <- scPagwas_main(
  Pagwas = NULL,
  gwas_data = GWAS_summ_example,
  add_eqtls = "OnlyTSS",
  block_annotation = gtf_df,
  Single_data = scRNAexample,
  Pathway_list = Genes_by_pathway_kegg,
  chrom_ld = chrom_ld
)
}
