% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Singlecell_heritability_contributions.R
\name{scPagwas_perform_regression}
\alias{scPagwas_perform_regression}
\title{scPagwas_perform_regression}
\usage{
scPagwas_perform_regression(Pagwas, Pathway_ld_gwas_data, ncores = 1)
}
\arguments{
\item{Pagwas}{Pagwas data list, default is "NULL"}

\item{Pathway_ld_gwas_data}{Pathway_ld_gwas_data}
}
\value{

}
\description{
Functions for inferring relevant annotations using the polyTest model.
}
\examples{
library(scPagwas)
scPagwas_perform_regression(Pagwas, ncores = ncores)
}
