% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Singlecell_heritability_contributions2.0.R
\name{scPagwas_perform_score}
\alias{scPagwas_perform_score}
\title{scPagwas_perform_score}
\usage{
scPagwas_perform_score(
  Pagwas,
  Pathway_ld_gwas_data,
  n.cores = 1,
  remove_outlier = TRUE,
  bignumber = 10000
)
}
\arguments{
\item{Pagwas}{pagwas}

\item{Pathway_ld_gwas_data}{Pagwas data list}

\item{n.cores}{cores}

\item{remove_outlier}{Whether to remove the outlier for scPagwas score.}

\item{bignumber}{the threshold of number of cells to change matrix to bigmemory data format, if you have enough
storage, you can set a large number to save time.}
}
\value{

}
\description{
Get the scPagwas score for each cells
}
\examples{
library(scPagwas)
scPagwas_perform_score(Pagwas)
}
