% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scPagwas_Visualization.R
\name{scPagwasPaHeritability_Visualization}
\alias{scPagwasPaHeritability_Visualization}
\title{scPagwasPaHeritability_Visualization}
\usage{
scPagwasPaHeritability_Visualization(
  Single_data = NULL,
  p_thre = 0.05,
  output.dirs = "test",
  FigureType = "tsne",
  width = 7,
  height = 7,
  lowColor = "#8479E1",
  midColor = "#F7F5F2",
  highColor = "#FD5D5D",
  p_color = "#E45826",
  size = 0.5,
  do_plot = F
)
}
\arguments{
\item{do_plot}{}
}
\value{

}
\description{
scPagwasPaHeritability_Visualization
}
