% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Singlecell_heritability_contributions.R
\name{scParameter_regression}
\alias{scParameter_regression}
\title{scParameter_regression}
\usage{
scParameter_regression(
  Pagwas_x,
  Pagwas_y,
  noise_per_snp,
  Rns,
  n.cores = 1,
  backingpath
)
}
\arguments{
\item{Pagwas_x}{x parameter for lm}

\item{Pagwas_y}{y parameter for lm}

\item{noise_per_snp}{noise}

\item{Rns}{the names for bk temfiles}

\item{n.cores}{cores for regression}

\item{backingpath}{file address for bk files, no "/" in the end.}
}
\description{
Find parameter estimates for the data.
In terms of settings, it can handle data from at least 10,000 cells in your computer.
}
