% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Link_pathway_blocks_gwas.R
\name{scPathway_block_splitfunc}
\alias{scPathway_block_splitfunc}
\title{Pathway_block_splitfunc}
\usage{
scPathway_block_splitfunc(
  pa_blocksnplist,
  subpca_scCell_mat,
  subdata_mat,
  rawPathway_list,
  snp_gene_df,
  ncores = 1
)
}
\arguments{
\item{ncores}{}
}
\value{

}
\description{
Pathway_block_splitfunc
}
