% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scPagwas-data.R
\docType{data}
\name{scRNAexample}
\alias{scRNAexample}
\title{scRNAexample}
\format{
a Seruat.
}
\source{
Generated from PBC
}
\usage{
data(scRNAexample)
}
\description{
A single cell example data from \code{
liver_data<-readRDS("/share/pub/qiuf/COVID/liver/Rpoly.rds")
#set Idents
liver_data@meta.data$annotation <- liver_data@meta.data$anno
Idents(liver_data) <- liver_data@meta.data$anno
#normalization
liver_data <- NormalizeData(liver_data, normalization.method = "LogNormalize", scale.factor = 10000)
#correct the cell names, need no specific signal
anno <- liver_data@meta.data$annotation
 anno<-str_replace_all(anno,"-",".")
 anno<-str_replace_all(anno," ",".")
anno[which(anno=="Inflammatory.monocyte/macrophages")]<-"Inflammatory.monocyte.macrophages"
Celltype_anno$annotation<-str_replace_all(anno,"\\+",".")
Idents(liver_data) <- as.factor(anno)
liver_data<-subset(liver_data,idents = c("Stellate.cell","unknow","γδ.T.cell"), invert = TRUE)
scRNAexample<-liver_data[,sample(ncol(liver_data),1000)]

 for (x in 1:length(unique(scRNAexample@meta.data$anno))){
 newnames <- paste0("Celltype",x)
 oldnames<-unique(scRNAexample@meta.data$anno)[x]
 scRNAexample@meta.data$anno[which(scRNAexample@meta.data$anno==oldnames)]<-newnames
 }
 Idents(scRNAexample)<-scRNAexample@meta.data$anno
 scRNAexample@meta.data<-scRNAexample@meta.data[,1:15]

 save(scRNAexample,file = "E:/RPakage/scPagwas/data/scRNAexample.RData")

}
}
\examples{
data(scRNAexample)
str(scRNAexample)
}
\keyword{datasets}
