% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Singlecell_heritability_contributions2.0.R
\name{SCxy2vector}
\alias{SCxy2vector}
\title{SCxy2vector}
\usage{
SCxy2vector(Pathway_ld_gwas_data = NULL)
}
\arguments{
\item{Pathway_ld_gwas_data}{the list of block information from Pagwas object}
}
\value{

}
\description{
Take a list of Pagwas - Pathway_ld_gwas_data and vectorize it.
}
