% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Single_data_input.R
\name{Single_data_input}
\alias{Single_data_input}
\title{Single_data_input}
\usage{
Single_data_input(
  Pagwas,
  Single_data,
  nfeatures = NULL,
  FilterSingleCell = FALSE,
  min.lib.size = 1000,
  min.reads = 10,
  min.detected = 5,
  min_clustercells = 50
)
}
\arguments{
\item{Pagwas}{Pagwas format}

\item{Single_data}{Input the Single data in seruat format, Idents should be
the celltypes annotation.}

\item{nfeatures}{The parameter for FindVariableFeatures,
NULL means select all genes}

\item{FilterSingleCell}{whther to filter the single cell data,if you
filter it before,choose FALSE, otherwise set TRUE.}

\item{min.lib.size}{Threshold for single data library.}

\item{min.reads}{Threshold for total reads fo each cells.}

\item{min.detected}{Threshold for total cells fo each gene.}

\item{min_clustercells}{Threshold for total cells fo each cluster.}
}
\value{
Pagwas
}
\description{
Input the Single data in seruat format
}
\examples{
library(scPagwas)
data(scRNAexample)
# Pagwas is better getting from GWAS_summary_input(),or NULL is right,
Pagwas <- Single_data_input(Pagwas = Pagwas, Single_data = scRNAexample)
}
