% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Snp2Gene.R
\name{Snp2Gene}
\alias{Snp2Gene}
\title{Snps to Genes}
\usage{
Snp2Gene(snp, refGene, marg = 10000)
}
\arguments{
\item{snp}{(char) PLINK bim file (only CHR, SNP, and BP columns
considered).}

\item{refGene}{(char) refseq table with header.}

\item{marg}{(integer) region upstream and downstream(default=0).}
}
\value{
data frame for snp mapping to gene.
}
\description{
Maps SNPs to their nearest genes within TSS windows.
}
\note{
updated GenomicRanges to version 1.30.1 on 01/30/2018
Requires ignore.strand=TRUE param to properly run distanceToNearest()
given unknown strand assignment for SNP array-based genotyping info
}
\examples{
library(scPagwas)
Pagwas <- list()
gwas_data <- bigreadr::fread2(system.file("extdata",
  "GWAS_summ_example.txt",
  package = "scPagwas"
))
Pagwas <- GWAS_summary_input(
  Pagwas = Pagwas,
  gwas_data = gwas_data,
  maf_filter = 0.1
)
Pagwas$snp_gene_df <- Snp2Gene(
  snp = Pagwas$gwas_data,
  refGene = block_annotation,
  marg = 10000
)
}
