% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SnpToGene.R
\name{SnpToGene}
\alias{SnpToGene}
\title{Functions for mapping Snps to Genes}
\usage{
SnpToGene(gwas_data, block_annotation, marg = 10000)
}
\arguments{
\item{gwas_data}{(data.frame) the result for GWAS_summary_input}

\item{block_annotation}{(data.frame) Start and end points for block
traits, usually genes.}

\item{marg}{(integer) region upstream and downstream(default=10000).}
}
\value{
data frame for gwas_data mapping to gene.
}
\description{
Maps SNPs to their nearest genes within TSS windows.
}
\examples{
library(scPagwas)
Pagwas <- list()
gwas_data <- bigreadr::fread(system.file("extdata",
  "GWAS_summ_example.txt",
  package = "scPagwas"
))
Pagwas <- GWAS_summary_input(
  Pagwas = Pagwas,
  gwas_data = gwas_data,
  maf_filter = 0.1
)
Pagwas$gwas_data_gene_df <- SnpToGene(
  gwas_data = Pagwas$gwas_data,
  block_annotation = block_annotation,
  marg = 10000
)
}
