% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Singlecell_heritability_contributions.R
\name{subget_Pathway_sclm}
\alias{subget_Pathway_sclm}
\title{subget_Pathway_sclm}
\usage{
subget_Pathway_sclm(
  path_block,
  subdata_mat,
  pathway,
  rawPathway_list,
  snp_gene_df,
  ncores = 1
)
}
\arguments{
\item{snp_gene_df}{}
}
\value{

}
\description{
subget_Pathway_sclm
}
