% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tissue_eqtls_Input.R
\name{Tissue_eqtls_Input}
\alias{Tissue_eqtls_Input}
\title{Tissue_eqtls_Input}
\usage{
Tissue_eqtls_Input(
  Pagwas = NULL,
  block_annotation = NULL,
  add_eqtls = "OnlyEqtls",
  eqtls_files = NULL,
  eqtl_p = 0.05,
  eqtls_cols = c("rs_id_dbSNP151_GRCh38p7", "variant_pos", "tss_distance", "gene_chr",
    "gene_start", "gene_end", "gene_name", "slope"),
  marg = 10000
)
}
\arguments{
\item{Pagwas}{Pagwas format data list inherit from other functions.}

\item{block_annotation}{gene annotation}

\item{add_eqtls}{There is options: "OnlyEqtls","Both"; "OnlyEqtls" means only eqtls snps will choose; "Both" means both eqtls and gene-TSS-window snps will choose;}

\item{eqtls_files}{eqtls files names,frome GTEx.}

\item{eqtl_p}{pvalue threshold for eqtls data.}

\item{eqtls_cols}{c("rs_id_dbSNP151_GRCh38p7","variant_pos","tss_distance","gene_chr", "gene_start", "gene_end","gene_name","slope")}

\item{marg}{gene-TSS-window size}
}
\value{

}
\description{
Get the eqtls data and interact with GWAS summary data.
}
\examples{
library(scPagwas)
data(eqtls_files)

Pagwas <- Tissue_eqtls_Input(Pagwas = Pagwas, add_eqtls = "OnlyEqtls", eqtls_files = eqtls_files)
}
