% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_functions.R
\name{utils_big_as.matrix}
\alias{utils_big_as.matrix}
\title{Convert a large sparse matrix into a dense matrix without errors}
\usage{
utils_big_as.matrix(sparseMat, n_slices_init = 1, verbose = T)
}
\arguments{
\item{sparseMat}{a big sparse matrix of a type coercible to dense Matrix::Matrix}

\item{n_slices_init}{initial number of slices. Default value 1, i.e. whole matrix}

\item{verbose}{print progress}
}
\value{
a dense matrix
}
\description{
https://rdrr.io/github/CBMR-Single-Cell-Omics-Platform/SCOPfunctions/src/R/utils.R
Avoid the following error
. Error in asMethod(object) :
. Cholmod error 'problem too large' at file ../Core/cholmod_dense.c,
. by slicing the matrix into submatrices, converting and cbinding them
. Increases number of slices until they succeed.
}
