% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Celltype_heritability_contributions.R
\name{xy2vector}
\alias{xy2vector}
\title{xy2vector}
\usage{
xy2vector(Pathway_ld_gwas_data = NULL)
}
\arguments{
\item{Pathway_ld_gwas_data}{the list of block information from Pagwas
object}
}
\value{
vector for x and y values.
}
\description{
Take a list of Pagwas - Pathway_ld_gwas_data and vectorize
it.
}
