% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scPagwas-data.R
\docType{data}
\name{block_annotation}
\alias{block_annotation}
\title{block_annotation}
\format{
a data.frame,
}
\source{
Generated from gencode.v44.annotation.gff3.gz
}
\usage{
data(block_annotation)
}
\description{
GRCh38 gencode.v44.annotation.gff3.gz
}
\examples{
data(block_annotation)
str(block_annotation)
}
\keyword{datasets}
