% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pagwas_perform_inference.R
\name{Boot_evaluate}
\alias{Boot_evaluate}
\title{Boot_evaluate}
\usage{
Boot_evaluate(Pagwas, bootstrap_iters = 200, part = 0.5, n.cores = 1)
}
\arguments{
\item{Pagwas}{Pagwas format, deault is NULL.}

\item{bootstrap_iters}{number of bootstrap iterations to run,default is 200}

\item{part}{number of bootstrap iterations to perform,default is 0.5}

\item{n.cores}{Parallel cores,default is 1. use detectCores() to check the cores in computer.}
}
\value{

}
\description{
Bootstrap to evaluate for confidence intervals.
}
