% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_result_visualize.R
\name{Bootstrap_estimate_Plot}
\alias{Bootstrap_estimate_Plot}
\title{Bootstrap_estimate_Plot}
\usage{
Bootstrap_estimate_Plot(
  Pagwas,
  figurenames = NULL,
  width = 9,
  height = 7,
  do_plot = F
)
}
\arguments{
\item{Pagwas}{Pagwas format of result from Pagwas_main()}

\item{figurenames}{The filename and address of the output plot,
default is "test_forest.pdf".IF figurenames= NULL,
only plot the figure and have not pdf figure.}

\item{width}{figure width}

\item{height}{figure height}

\item{do_plot}{whether to plot the plot.}
}
\value{
A forest plot with the table of p values
}
\description{
This forest plot shows the correct estimate values and
95\% CI for different celltyppes, using the ggplot packages
}
\examples{
library(scPagwas)
# Pagwas is the result of Pagwas_main()
Bootstrap_estimate_Plot(Pagwas=Pagwas,
                        figurenames = "test_forest.pdf",
                        width = 9, height = 7)
}
