% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_result_visualize.R
\name{bootstrap_P_Barplot}
\alias{bootstrap_P_Barplot}
\title{bootstrap_P_Barplot}
\usage{
bootstrap_P_Barplot(
  Pagwas,
  title = "Test scPagwas",
  file = "test_barplot.pdf",
  width = 5,
  height = 7
)
}
\arguments{
\item{Pagwas}{Pagwas format of result}

\item{title}{the title names of the plot}

\item{file}{the filename and address of the output plot,default is "test_barplot.pdf"}

\item{width}{figure width, default is 5}

\item{height}{figure height,default is 7}
}
\value{

}
\description{
bootstrap_P_Barplot
}
\examples{
library(scPagwas)
bootstrap_P_Barplot(Pagwas,title="Test scPagwas",file="test_barplot.pdf",width=5,height =7)
}
