% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_result_visualize.R
\name{Bootstrap_P_Barplot}
\alias{Bootstrap_P_Barplot}
\title{Bootstrap_P_Barplot}
\usage{
Bootstrap_P_Barplot(
  Pagwas,
  title = "Test scPagwas",
  file = "test_barplot.pdf",
  width = 5,
  height = 7
)
}
\arguments{
\item{Pagwas}{Pagwas format of result in Pagwas_main()}

\item{title}{The title names of the plot}

\item{file}{The filename and address of the output plot,default is "test_barplot.pdf"}

\item{width}{figure width, default is 5}

\item{height}{figure height,default is 7}
}
\value{
A figure of barplot in pdf format, red color is significant.
}
\description{
This bar plot shows the -log2(p value) for bootstrap result, using the ggplot packages
}
\examples{
library(scPagwas)
# Pagwas is the result of Pagwas_main()
Bootstrap_P_Barplot(Pagwas, title = "Test scPagwas", file = "test_barplot.pdf", width = 5, height = 7)
}
