% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Celltype_heritability_contributions.R
\name{Celltype_heritability_contributions}
\alias{Celltype_heritability_contributions}
\title{Celltype_heritability_contributions}
\usage{
Celltype_heritability_contributions(Pagwas, iters = 200, part = 0.5)
}
\arguments{
\item{Pagwas}{Pagwas format, deault is NULL.}

\item{iters}{number of bootstrap iterations to perform}

\item{part}{number of bootstrap iterations to perform,default is 0.5}

\item{ncores}{Parallel cores,default is 1. use detectCores() to check the cores in computer.}
}
\value{

}
\description{
Celltype_heritability_contributions
}
