% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scPagwas-data.R
\docType{data}
\name{chrom_ld}
\alias{chrom_ld}
\title{ld_folder
use vcftools to get ped and map files
LD data from \code{
/share/apps/vcftools/bin/vcftools --vcf /share/pub/dengcy/Singlecell/COVID19/1000genomes_all_genotypes.vcf --plink-tped --out /share/pub/dengcy/Singlecell/COVID19/1000genomes_all_genotypes}
\format{
list
}
\source{
Generated from PLINK 1.90 linux
}
\description{
/share/pub/dengcy/Singlecell/COVID19/PLINK/plink --tfile /share/pub/dengcy/Singlecell/COVID19/1000genomes_all_genotypes --recode --out /share/pub/dengcy/Singlecell/COVID19/1000genomes_all_genotypes
/share/pub/dengcy/Singlecell/COVID19/PLINK/plink --map /share/pub/dengcy/Singlecell/COVID19/1000genomes_all_genotypes.map --ped /share/pub/dengcy/Singlecell/COVID19/1000genomes_all_genotypes.ped --allow-no-sex --autosome --r2 --ld-window-kb 1000 --ld-window-r2 0.2 --out /share/pub/dengcy/Singlecell/COVID19/ld_1000genome
covid_ld<-read.delim("/share/pub/dengcy/Singlecell/COVID19/ld_1000genome.ld")
covid_ld<-covid_ld\link{!(covid_ld$\%in\% 23),}
colnames(covid_ld)\link{7}<-"R"
lapply(unique(covid_ld$CHR_A), function(i){
a<-data.table(covid_ld\link{covid_ld$CHR_A == i,})
file_name <- paste0("/share/pub/dengcy/Singlecell/COVID19/data/LD/",i,".Rds")
saveRDS(a, file = file_name)
})
}
\details{

}
\examples{
data(chrom_ld)
str(chrom_ld)
}
