% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_functions.R
\name{corSparse}
\alias{corSparse}
\title{various association measures between sparse matrices}
\usage{
corSparse(X, Y)
}
\arguments{
\item{X}{matrix}

\item{Y}{matrix or vector}
}
\value{

}
\description{
Pearson correlation matrix between columns of X, Y
modification the code from:
http://stackoverflow.com/questions/5888287/running-cor-or-any-variant-over-a-sparse-matrix-in-r
https://github.com/cysouw/qlcMatrix/blob/master/R/assoc.R
}
\note{
Note that results larger than 1e4 x 1e4 will become very slow, because the resulting matrix is not sparse anymore.
}
