% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scPagwas_Visualization.R
\name{fortify.Seurat.tsne}
\alias{fortify.Seurat.tsne}
\title{fortify.Seurat.tsne}
\usage{
\method{fortify}{Seurat.tsne}(x)
}
\arguments{
\item{x}{seruat}
}
\description{
set data frame to ggplot
}
