% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pagwas_perform_inference.R
\name{Get_Pathway_heritability_contributions}
\alias{Get_Pathway_heritability_contributions}
\title{Get_Pathway_heritability_contributions}
\usage{
Get_Pathway_heritability_contributions(pca_cell_df, parameters)
}
\arguments{
\item{pca_cell_df}{pca score dataframe}

\item{parameters}{parameter fit}
}
\value{

}
\description{
Caclulate predicted block values based on block information and model fit.
}
