% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Singlecell_heritability_contributions.R
\name{get_Pathway_sclm}
\alias{get_Pathway_sclm}
\title{get_Pathway_sclm}
\usage{
get_Pathway_sclm(Pagwas, path_block, pathway, ncores = 1)
}
\arguments{
\item{path_block}{Element in Pagwas}

\item{ncores}{}
}
\value{

}
\description{
get_Pathway_sclm
}
