% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Singlecell_heritability_contributions.R
\name{get_Pathway_sclm}
\alias{get_Pathway_sclm}
\title{get_Pathway_sclm}
\usage{
get_Pathway_sclm(
  Pathway_ld_gwas_data,
  pca_scCell_mat,
  data_mat,
  rawPathway_list,
  snp_gene_df,
  n.cores = 1
)
}
\arguments{
\item{Pathway_ld_gwas_data}{Element in Pagwas}

\item{pca_scCell_mat}{Element in Pagwas}

\item{data_mat}{Element in Pagwas}

\item{rawPathway_list}{Element in Pagwas}

\item{snp_gene_df}{Element in Pagwas}

\item{n.cores}{n.cores}

\item{split_n}{number of times to compute the singlecell result}
}
\value{

}
\description{
get_Pathway_sclm
}
