% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWAS_summary_input.R
\name{GWAS_summary_input}
\alias{GWAS_summary_input}
\title{GWAS_summary_input}
\usage{
GWAS_summary_input(
  Pagwas = NULL,
  gwas_data = NULL,
  maf_filter = 0.01,
  Sex_filter = TRUE,
  MHC_filter = TRUE,
  gwas_z_filter = -1
)
}
\arguments{
\item{Pagwas}{Pagwas format, deault is NULL.}

\item{gwas_data}{A data set object, need a data frame format, There must be have colomue with 'chrom', 'pos', 'rsid', 'beta', 'se', 'maf'}

\item{maf_filter}{Filter the maf, default is <0.01}

\item{Sex_filter}{Filter the SNPs in Sex chromosome,  default is TRUE.}

\item{MHC_filter}{Filter the SNPs in MHC chromosome,  default is TRUE.}

\item{gwas_z_filter}{Filter the z-score, calculated by abs(beta/se)}
}
\value{

}
\description{
Input and progress the GWAS summary statistics data
filter the maf and the sex chrom data
}
\examples{
library(scPagwas)
data(GWAS_summ_example)
Pagwas <- GWAS_summary_input(Pagwas = NULL, gwas_data = GWAS_summ_example)
}
