% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heritability_cor_scatterplot.R
\name{heritability_cor_scatterplot}
\alias{heritability_cor_scatterplot}
\title{heritability_cor_scatterplot}
\usage{
heritability_cor_scatterplot(
  gene_heri_cor = NULL,
  topn_genes_label = 10,
  color_low = "#035397",
  color_high = "#F32424",
  color_mid = "white",
  text_size = 3,
  do_plot = TRUE,
  max.overlaps = 10,
  figurenames = NULL,
  width = 7,
  height = 7
)
}
\arguments{
\item{gene_heri_cor}{PCC result for scPagwas.}

\item{topn_genes_label}{10, the number of top genes for correlation.}

\item{color_low}{low color}

\item{color_high}{high color}

\item{color_mid}{mid color}

\item{text_size}{text size}

\item{do_plot}{parametr for "geom_label_repel", Exclude text labels that overlap too many things. Defaults to 10.}

\item{max.overlaps}{mac over}

\item{figurenames}{The filename and address of the output plot}

\item{width}{figure width}

\item{height}{figure height}
}
\value{
scatterplot
}
\description{
Plot the scatterplot for the pearson correlation coefficients(PCC) for each genes.
pearson methods for each gene expression and gPAS score in each cells.
}
\examples{
load(system.file("extdata", "Pagwas_data.RData", package = "scPagwas"))
heritability_cor_scatterplot(
  gene_heri_cor = Pagwas_data@misc$PCC,
  topn_genes_label = 10,
  color_low = "#035397",
  color_high = "#F32424",
  color_mid = "white",
  text_size = 2,
  do_plot = TRUE,
  max.overlaps = 20,
  width = 7,
  height = 7
)
}
\author{
Chunyu Deng
}
\keyword{PCC}
\keyword{each}
\keyword{for}
\keyword{genes.}
\keyword{heritability_cor_scatterplot,}
\keyword{plot}
\keyword{scatterplot}
\keyword{the}
