% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Link_pathway_blocks_gwas.R
\name{Link_pathway_blocks_gwas}
\alias{Link_pathway_blocks_gwas}
\title{Link_pathway_blocks_gwas}
\usage{
Link_pathway_blocks_gwas(Pagwas, chrom_ld = NULL, n.cores = 1)
}
\arguments{
\item{Pagwas}{Pagwas format, deault is NULL.}

\item{chrom_ld}{LD data for 22 chromosome.}

\item{n.cores}{Parallel cores,default is 1. use detectCores()
to check the cores in computer.}
}
\value{

}
\description{
Link pathway blocks region and gwas summary snps
Takes block information, potentially independent LD pathway_blocks
or gene blocks,
}
\examples{
library(scPagwas)
data(chrom_ld)
Pagwas <- Link_pathway_blocks_gwas(Pagwas = Pagwas, chrom_ld = chrom_ld)
}
