% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Celltype_heritability_contributions.R
\name{link_pwpca_block}
\alias{link_pwpca_block}
\title{link_pwpca_block}
\usage{
link_pwpca_block(Pagwas)
}
\arguments{
\item{Pagwas}{Pagwas format, deault is NULL.}
}
\value{

}
\description{
Link the pca score and expression for each pathway genes
for each block
Requires rownames that are identitcal to block labels loaded previously.
}
\examples{
library(scPagwas)
# Pagwas should have inhibit data
Pagwas <- link_pwpca_block(Pagwas)
}
