% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_scCell_pwpca_block.R
\name{link_scCell_pwpca_block}
\alias{link_scCell_pwpca_block}
\title{link_scCell_pwpca_block}
\usage{
link_scCell_pwpca_block(
  Pagwas,
  n.cores = 1,
  scPagwasSession = "scPagwasSession"
)
}
\arguments{
\item{Pagwas}{Pagwas format, deault is NULL.}

\item{n.cores}{Parallel cores,default is 1. use detectCores() to
check the cores in computer.}

\item{scPagwasSession}{"scPagwasSession"}
}
\value{

}
\description{
link the single cell pca score and expression for each
pathway genes for each block(single cell)
construct the final matrix for regression in blocks
}
\examples{
library(scPagwas)
}
