% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scPagwas_main.R
\name{merge_pagwas}
\alias{merge_pagwas}
\title{merge_pagwas}
\usage{
merge_pagwas(
  Single_data = NULL,
  n_topgenes = 1000,
  output.dirs = "splice_scPagwas",
  assay = "RNA",
  random = FALSE,
  seed = 1234,
  random_times = 10,
  proportion = 0.3
)
}
\arguments{
\item{Single_data}{The Whole single cell data set. rds files in seurat format.
The filesnames or seurat format data.}

\item{n_topgenes}{(integr)Number of top associated gene selected to
calculate the scPagwas score;}

\item{output.dirs}{The files names for the result of scPagwas_main function.}

\item{assay}{The assay used for single cell data.}

\item{random}{Whether split the Single_data in random.
Using id when the Single_data is too big to run.}

\item{seed}{The seed for random.}

\item{random_times}{The times for random, default is 10.}

\item{proportion}{The proportion for spliting the Single_data; the bigger for
Single_data the less proportion for spliting, and the more times for random_times.}
}
\value{
seurat format result for scPagwas.
}
\description{
Merging several pagwas result to one result.When your single
cell data is too large to run, you can split your data into several
split, then use this function to merge them. Split the data randomly or
in order of cellytpes.
}
\note{
1.The rownames genes and gwas data must uniform.
2.The merge_pagwas function only can use to merge the seurat format
result.
That is "singlecell = T;celltype =F" can use to all the split data,
"singlecell = T;celltype = T" can be used only when the data is splited
by celltypes.
but "singlecell = F;celltype =T" can not be used.
}
\examples{
library(Seurat)
scRNAexample <-readRDS(system.file("extdata", "scRNAexample.rds", package = "scPagwas"))
#set the output.dirs for any directory.
set.dirs="D:/OneDrive/GWAS_Multiomics/Pagwas/test/"
#set the number of split time, it depend on the size of your single cell data.
#There have two form of spliting the data:

#########
##1.By Random
#Create the random index number.
n_split=2
Split_index <- rep(1:n_split, time = ceiling(ncol(scRNAexample)/n_split), length = ncol(scRNAexample))

for (i in 1:n_split) {
  Example_splice <- scRNAexample[,Split_index==i]
  saveRDS(Example_splice,file = paste0(set.dirs,"Example_splice",i,".rds"))
}

#########
###1.By celltypes
#Create the random index number.
for (i in levels(Idents(scRNAexample))) {
  Example_splice <- subset(scRNAexample,idents = i)
  saveRDS(Example_splice,file = paste0(set.dirs,"Example_splice",i,".rds"))
}
Pagwas<-list()
gwas_data <- bigreadr::fread2(system.file("extdata", "GWAS_summ_example.txt", package = "scPagwas"))
Pagwas <- GWAS_summary_input(
  Pagwas = Pagwas,
  gwas_data = gwas_data,
  maf_filter = 0.1
)
Pagwas$snp_gene_df <- Snp2Gene(snp = Pagwas$gwas_data, refGene = block_annotation, marg = 10000)
names(Pagwas)
setwd(set.dirs)

for (i in 1:n_split) {
  scPagwas_main(Pagwas =Pagwas,
                gwas_data =NULL,
                Single_data = paste0("Example_splice",i,".rds"), #the
                output.prefix=i, #the prefix can be the circulation coefficient i.
                output.dirs="splice_scPagwas", #the output.dirs shoukd be the same for each circulation
                Pathway_list=Genes_by_pathway_kegg,
                run_split=TRUE, #This parameter is set to run single result for each split result.
                ncores=1,
                min_clustercells=1, #the minimum cluster cell number should be 1.
                assay="RNA",
                block_annotation = block_annotation,
                chrom_ld = chrom_ld)
}
if(length(SOAR::Objects())>0){
  SOAR::Remove(SOAR::Objects())
}

setwd(set.dirs) #you must set the workfiles the same as before
Pagwas_integrate <- merge_pagwas(Single_data = system.file("extdata", "scRNAexample.rds", package = "scPagwas") ,# read the whole single cell data.
                                output.dirs="splice_scPagwas", #The same with scPagwas_main functionn_topgenes = 1000,
                                assay='RNA',
                                random=T,
                                seed=1234,
                                random_times=10,
                                proportion=0.3)


}
\author{
Chunyu Deng
}
\keyword{integrate}
\keyword{merge_pagwas,}
\keyword{result.}
\keyword{scPagwas}
\keyword{spliced}
\keyword{the}
