% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Celltype_heritability_contributions.R
\name{Pagwas_perform_regression}
\alias{Pagwas_perform_regression}
\title{Pagwas_perform_regression}
\usage{
Pagwas_perform_regression(Pathway_ld_gwas_data)
}
\arguments{
\item{Pathway_ld_gwas_data}{Pagwas format, deault is NULL.}
}
\value{

}
\description{
Run regression
}
