% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pagwas_perform_inference.R
\name{Pagwas_perform_regression}
\alias{Pagwas_perform_regression}
\title{Pagwas_perform_regression}
\usage{
Pagwas_perform_regression(Pagwas, iters = iters, part = 0.5, n.cores = n.cores)
}
\arguments{
\item{Pagwas}{Pagwas format, deault is NULL.}

\item{iters}{number of bootstrap iterations to perform}

\item{part}{number of bootstrap iterations to perform,default is 0.5}

\item{n.cores}{Parallel cores,default is 1. use detectCores() to check the cores in computer.}
}
\value{

}
\description{
Run regression
}
\examples{
library(scPagwas)
Pagwas <- Pagwas_perform_regression(Pagwas, iters = 200)

}
