% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pagwas_result_integrate.r
\name{Pagwas_result_integtate}
\alias{Pagwas_result_integtate}
\title{Pagwas_result_integtate}
\usage{
Pagwas_result_integtate(
  Pagwas,
  seruat_return,
  output.dirs,
  output.prefix,
  n_topgenes,
  assay
)
}
\arguments{
\item{assay}{}
}
\value{

}
\description{
Pagwas_result_integtate
}
