% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_functions.R
\name{papply}
\alias{papply}
\title{papply}
\usage{
papply(..., n.cores = detectCores())
}
\arguments{
\item{n.cores}{default is 1.}

\item{...parameters}{to pass to lapply, mclapply, bplapply, etc.}
}
\value{

}
\description{
wrapper around different mclapply mechanisms
Abstracts out mclapply implementation, and defaults to lapply when only one core is requested (helps with debugging)
}
