% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pathway_annotation_input.R
\name{Pathway_annotation_input}
\alias{Pathway_annotation_input}
\title{Pathway_annotation_input}
\usage{
Pathway_annotation_input(Pagwas, n.cores = 1)
}
\arguments{
\item{Pagwas}{Pagwas format data list inherit from other functions.}

\item{n.cores}{Parallel cores,default is 1. use detectCores() to check the cores in computer.}
}
\value{

}
\description{
Deal with the pathway genes with snp blocks
}
\examples{
library(scPagwas)
data(Genes_by_pathway.kegg)
#the Pagwas should be after Single_data_input() and GWAS_summary_input(),Pathway_pcascore_run()
Pagwas<-Pathway_annotation_input(Pagwas=Pagwas)

}
