% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Link_pathway_blocks_gwas.R
\name{Pathway_block_func}
\alias{Pathway_block_func}
\title{Pathway_block_func}
\usage{
Pathway_block_func(
  Pagwas = NULL,
  Pachrom_block_list,
  chrom_gwas_list,
  ncores = 1,
  singlecell = T,
  celltype = T
)
}
\arguments{
\item{Pagwas}{NULL}

\item{Pachrom_block_list}{result for chrome block}

\item{ncores}{cores}
}
\value{

}
\description{
Pathway_block_func
}
