% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Link_pathway_blocks_gwas.R
\name{Pathway_block_func}
\alias{Pathway_block_func}
\title{Pathway_block_func}
\usage{
Pathway_block_func(
  Pagwas = NULL,
  Pachrom_block_list,
  chrom_gwas_list,
  singlecell = T,
  celltype = T,
  chrom_ld
)
}
\arguments{
\item{Pagwas}{NULL}

\item{Pachrom_block_list}{result for chrome block}

\item{chrom_gwas_list}{list of gwas block in chrom}

\item{singlecell}{whether to run singlecell progress}

\item{celltype}{whether to run celltype progress}

\item{chrom_ld}{default is block_annotation}
}
\value{
Pagwas result list including Pathway block.
}
\description{
Pathway_block_func
}
