% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pathway_pcascore_run.R
\name{Pathway_pcascore_run}
\alias{Pathway_pcascore_run}
\title{Pathway_pcascore_run
The function can calculate the pathway pcascore for celltypes and single cells.}
\usage{
Pathway_pcascore_run(
  Pagwas = NULL,
  Pathway_list = NULL,
  n.cores = 1,
  min.pathway.size = 10,
  max.pathway.size = 1000
)
}
\arguments{
\item{Pagwas}{Pagwas format}

\item{Pathway_list}{Pathawy gene list(gene symbol),list name is pathway name.}

\item{n.cores}{Parallel cores,default is 1. use detectCores() to check the cores in computer.}

\item{min.pathway.size}{Threshold for min pathway size.}

\item{max.pathway.size}{Threshold for max pathway size.}
}
\value{
Pagwas
}
\description{
Pathway_pcascore_run
The function can calculate the pathway pcascore for celltypes and single cells.
}
\examples{
library(scPagwas)
data(Genes_by_pathway.kegg)
#the Pagwas shoul be after Single_data_input() and GWAS_summary_input()
Pagwas<-Pathway_pcascore_run(Pagwas=Pagwas,Pathway_list=Genes_by_pathway.kegg)

}
