% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pathway_pcascore_run.R
\name{PathwayPCAtest}
\alias{PathwayPCAtest}
\title{PathwayPCAtest}
\usage{
PathwayPCAtest(Pathway_list, scCounts)
}
\arguments{
\item{Pathway_list}{Pathawy gene list(gene symbol),list name is pathway name.}

\item{scCounts}{single cell Counts pathway}
}
\value{
list of pca score

library(scPagwas)
Pathway_list<-
}
\description{
Calculate pca score
}
