% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bar_positie_nagtive.R
\name{plot_bar_positie_nagtive}
\alias{plot_bar_positie_nagtive}
\title{plot_bar_positie_nagtive}
\usage{
plot_bar_positie_nagtive(
  seurat_obj,
  var_ident,
  var_group,
  vec_group_colors = NULL,
  f_color = colorRampPalette(brewer.pal(n = 11, name = "RdYlBu")),
  do_plot = F,
  title = NULL,
  fontsize_title = 24,
  fontsize_axistitle_x = 18,
  fontsize_axistitle_y = 18,
  fontsize_axistext_x = 12,
  fontsize_axistext_y = 12,
  fontsize_legendtitle = 12,
  fontsize_legendtext = 10,
  aspect.ratio = 1.2,
  figurenames = NULL,
  width = 7,
  height = 7
)
}
\arguments{
\item{seurat_obj}{Seurat object (Seurat ^3.0)}

\item{var_ident}{the identify variable, character}

\item{var_group}{the group variable, character}

\item{vec_group_colors}{a vector of colors, named by corresponding group. Length must match number of groups. Character}

\item{f_color}{if vec_group_colors is not provided, the user may instead provide a function f_color() that takes as its only argument the number of colors to generate}

\item{do_plot}{Whether to plot, logical}

\item{title}{NULL to leave out}

\item{fontsize_title}{NULL to leave out}

\item{fontsize_axistitle_x}{NULL to leave out}

\item{fontsize_axistitle_y}{NULL to leave out}

\item{fontsize_axistext_x}{NULL to leave out}

\item{fontsize_axistext_y}{NULL to leave out}

\item{fontsize_legendtitle}{NULL to leave out}

\item{fontsize_legendtext}{NULL to leave out}

\item{figurenames}{The filename and address of the output plot,}

\item{width}{figure width}

\item{height}{figure height}
}
\value{
a ggplot2 object
}
\description{
Generate barplot of identity group composition
Thanks to the SCOPfunctions package
https://github.com/CBMR-Single-Cell-Omics-Platform/SCOPfunctions/blob/main/R/plot.R
Generate a percentage barplot that shows the composition of each identity
(e.g. sample)in terms of groups (e.g. positive and negative cells for scPagwas)
}
\examples{
p <- plot_barIdentGroup(seurat_obj=seu, var_ident="sample",var_group="cluster")
}
