% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Pathway_contribution_Network.r
\name{plot_network}
\alias{plot_network}
\title{Thanks to the SCOPfunctions package "https://github.com/CBMR-Single-Cell-Omics-Platform/SCOPfunctions"}
\usage{
plot_network(
  mat_datExpr,
  vec_geneImportance,
  vec_genes_highlight = c(),
  n_max_genes = 50,
  igraph_algorithm = "drl",
  fontface_labels = "bold.italic",
  color_edge = "grey70",
  fontSize_label_lg = 1,
  fontSize_legend_lg = 1,
  fontSize_legend_xlg = 1,
  edge_thickness = 1,
  do_plot = TRUE
)
}
\description{
Thanks to the SCOPfunctions package "https://github.com/CBMR-Single-Cell-Omics-Platform/SCOPfunctions"
}
