% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_vln_Corgenes.R
\name{plot_vln_Corgenes}
\alias{plot_vln_Corgenes}
\title{plot_vln_Corgenes}
\usage{
plot_vln_Corgenes(
  seurat_obj,
  assay,
  slot,
  var_group,
  vec_features,
  vec_group_colors = NULL,
  f_color = colorRampPalette(brewer.pal(n = 11, name = "RdYlBu")),
  flip = T,
  do_plot = F,
  pt.size = 0,
  feature_fontface = "bold.italic",
  fontsize_axistext_x = 12,
  fontsize_axistext_y = 12,
  aspect.ratio = NULL,
  figurenames = NULL,
  width = 7,
  height = 7
)
}
\arguments{
\item{seurat_obj}{Seurat object (Seurat ^3.0)}

\item{assay}{seurat_obj assay to use}

\item{slot}{seurat_obj slot to use}

\item{var_group}{the group variable, character}

\item{vec_features}{a vector of features to plot in the violin plot}

\item{vec_group_colors}{a vector of colors, named by corresponding group. Length must match number of groups. Character}

\item{f_color}{if vec_group_colors is not provided, the user may instead provide a function f_color() that takes as its only argument the number of colors to generate}

\item{flip}{if TRUE (default), groups are rows and features are columns, and vice-verso for FALSE}

\item{do_plot}{Whether to plot, logical}

\item{pt.size}{size of jitter in the violin plots. Set to 0 (default) to omit}

\item{figurenames}{The filename and address of the output plot,}

\item{width}{figure width}

\item{height}{figure height}
}
\value{
ggplot2 object
}
\description{
Make grid of violin plots
Thanks to the SCOPfunctions package
https://github.com/CBMR-Single-Cell-Omics-Platform/SCOPfunctions/blob/main/R/plot.R

produce a n_celltype * n_genes grid of violin plots
}
\examples{
top5genes <- rownames(Pagwas$gene_heritability_correlation)[order(Pagwas$gene_heritability_correlation, decreasing = T)[1:5]]
library("RColorBrewer")
library("ggplot2")

plot_vln_Corgenes(
  seurat_obj = scRNAexample,
  assay = "RNA", slot = "data",
  var_group = "anno", # 缁嗚優cluster娉ㄩ噴鍒㤼㸷
  vec_features = top5genes,
  vec_group_colors = pal_d3(alpha = 0.5)(10)
)
}
