% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redunce_pathway.R
\name{reduce_pathway}
\alias{reduce_pathway}
\title{reduce_pathway}
\usage{
reduce_pathway(
  pathway_seed = names(Genes_by_pathway_kegg)[sample(1:length(Genes_by_pathway_kegg),
    20)],
  pathway_list = Genes_by_pathway_kegg,
  remove_proporion = 0.7
)
}
\arguments{
\item{remove_proporion}{}
}
\value{

}
\description{
reduce_pathway
}
