% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redunce_pathway.R
\name{reduce_pathway}
\alias{reduce_pathway}
\title{reduce_pathway}
\usage{
reduce_pathway(
 
    pathway_seed = names(Genes_by_pathway_kegg)[sample(seq_len(length(Genes_by_pathway_kegg)),
    20)],
  pathway_list = Genes_by_pathway_kegg,
  remove_proporion = 0.7
)
}
\arguments{
\item{pathway_seed}{randome select the seed pathways, sometimes can be the
unredundant pathway set.}

\item{pathway_list}{the list of pathway need to reduce.}

\item{remove_proporion}{the propotion of pathways after reduce.}
}
\value{

}
\description{
reduce_pathway
}
\examples{
reduce_pathway(pathway_seed = names(Genes_by_pathway_kegg)[sample(1:length(Genes_by_pathway_kegg), 20)],
                           pathway_list = Genes_by_pathway_kegg,
                           remove_proporion = 0.7)
}
