% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Singlecell_heritability_contributions.R
\name{scBoot_evaluate}
\alias{scBoot_evaluate}
\title{scBoot_evaluate}
\usage{
scBoot_evaluate(
  Pagwas,
  Pathway_ld_gwas_data,
  bootstrap_iters,
  ncores = 1,
  part = 0.5
)
}
\arguments{
\item{Pagwas}{Pagwas object}

\item{bootstrap_iters}{number of bootstrap iterations to run}

\item{part}{number of bootstrap iterations to perform,default is 0.5}
}
\value{

}
\description{
Bootstrap to evaluate for confidence intervals.
}
