% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Singlecell_heritability_contributions.R
\name{scGet_gene_heritability_correlation}
\alias{scGet_gene_heritability_correlation}
\title{scGet_gene_heritability_contributions}
\usage{
scGet_gene_heritability_correlation(scPagwas.gPAS.score, data_mat)
}
\arguments{
\item{scPagwas.gPAS.score}{score of scPagwas pathway lm}

\item{data_mat}{matrix for single cell data}
}
\value{
result list including gene heritability correlation
}
\description{
scGet_gene_heritability_contributions
}
