% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Singlecell_heritability_contributions.R
\name{scGet_gene_heritability_correlation}
\alias{scGet_gene_heritability_correlation}
\title{scGet_gene_heritability_contributions}
\usage{
scGet_gene_heritability_correlation(Pagwas)
}
\arguments{
\item{Pagwas}{result of scPagwas}
}
\value{
Pagwas result list including gene heritability correlation
}
\description{
scGet_gene_heritability_contributions
}
