% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Singlecell_heritability_contributions2.0.R
\name{scGet_gene_heritability_correlation}
\alias{scGet_gene_heritability_correlation}
\title{scGet_gene_heritability_contributions}
\usage{
scGet_gene_heritability_correlation(Pagwas)
}
\arguments{
\item{Pagwas}{result of scPagwas}
}
\value{

}
\description{
scGet_gene_heritability_contributions
}
\examples{
Single_data<-readRDS("E:/RPakage/scPagwas/inst/extdata/scRNAexample.rds")
Pagwas$sparse_cor<-scGet_gene_heritability_contributions(
scPagwas_score=Pagwas$scPagwas_score,
data_mat=Seurat::GetAssayData(object = Single_data[,names(Pagwas$scPagwas_score)], slot = "data"))
}
