% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scPagwas_perform_score.R
\name{scGet_gene_heritability_correlation}
\alias{scGet_gene_heritability_correlation}
\title{scGet_gene_heritability_contributions}
\usage{
scGet_gene_heritability_correlation(scPagwas_score, data_mat)
}
\arguments{
\item{scPagwas_score}{result of scPagwas}

\item{data_mat}{the data matrix from single cell data}
}
\value{

}
\description{
scGet_gene_heritability_contributions
}
