% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scPagwas_main.R
\name{scPagwas_main}
\alias{scPagwas_main}
\title{Main wrapper functions for scPagwas}
\usage{
scPagwas_main(
  Pagwas = NULL,
  gwas_data = NULL,
  output.prefix,
  add_eqtls = c("OnlyTSS", "OnlyEqtls", "Both"),
  eqtls_files = NULL,
  eqtls_cols = c("rs_id_dbSNP151_GRCh38p7", "variant_pos", "tss_distance", "gene_chr",
    "gene_start", "gene_end", "gene_name", "pval_beta"),
  block_annotation = NULL,
  Single_data = NULL,
  assay = c("RNA", "SCT"),
  nfeatures = NULL,
  Pathway_list = NULL,
  chrom_ld = NULL,
  split_n = 1,
  marg = 10000,
  maf_filter = 0.01,
  min_clustercells = 10,
  min.pathway.size = 5,
  max.pathway.size = 300,
  param.file = T,
  remove_outlier = T,
  SimpleResult = F,
  log.file = "scPagwas.run.log",
  n.cores = 1
)
}
\arguments{
\item{Pagwas}{(list)Pagwas data list, default is "NULL"}

\item{gwas_data}{(data.frame)GWAS summary data}

\item{output.prefix}{output.prefix,prefix used for output tables}

\item{add_eqtls}{(character)There are three options: "OnlyTSS","OnlyEqtls","Both"; "OnlyTSS" means there is only snps within TSS window;"OnlyEqtls" represents only significant snps in eqtls files;  "Both" represents both significant snps in eqtls and the other snps in GWAS summary files;}

\item{eqtls_files}{(character)eqtls files names,frome GTEx.}

\item{eqtls_cols}{(character) the needed columns in eqtls_files; In our examples files, the columns including: c("rs_id_dbSNP151_GRCh38p7","variant_pos","tss_distance","gene_chr", "gene_start", "gene_end","gene_name")}

\item{block_annotation}{(data.frame) Start and end points for block traits, usually genes.}

\item{Single_data}{(Seruat)Input the Single data in seruat format, Idents should be the celltypes annotation.}

\item{assay}{(character)assay data of your single cell data to use,default is "RNA"}

\item{nfeatures}{(integr) The parameter for FindVariableFeatures, NULL means select all genes}

\item{Pathway_list}{(list,character) pathway gene sets list}

\item{chrom_ld}{(list,numeric)LD data for 22 chromosome.}

\item{split_n}{number of times to compute the singlecell result}

\item{marg}{(integr)gene-TSS-window size}

\item{maf_filter}{(numeric)Filter the maf, default is 0.01}

\item{min_clustercells}{(integr)Only use is when FilterSingleCell is TRUE.Threshold for total cells fo each cluster.default is 10}

\item{min.pathway.size}{(integr)Threshold for min pathway gene size. default is 5}

\item{max.pathway.size}{(integr)Threshold for max pathway gene size. default is 300}

\item{param.file}{(logical)whether save parameters used for scPagwas.}

\item{remove_outlier}{(logical)Whether to remove the outlier for scPagwas score.}

\item{SimpleResult}{(logical)whether simplify the scPagwas result.}

\item{log.file}{(character)log.file}

\item{n.cores}{(integr)Parallel cores,default is 1. use detectCores() to check the cores in computer.}
}
\value{

}
\description{
Main Pagwas wrapper re-progress function.
}
\details{
The entry point for Pagwas analysis.
}
\examples{
library(scPagwas)
data(Genes_by_pathway_kegg)
data(GWAS_summ_example)
data(gtf_df)
data(scRNAexample)
#data(eqtls_files)
data(chrom_ld)
#1. OnlyTSS
Pagwas<-scPagwas_main(Pagwas = NULL,
                    gwas_data = GWAS_summ_example,
                    add_eqtls="OnlyTSS",
                    block_annotation = gtf_df,
                    Single_data = scRNAexample,
                    Pathway_list=Genes_by_pathway_kegg,
                    chrom_ld = chrom_ld)

}
