% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scPagwas_perform_inference.R
\name{scPagwas_perform_regression}
\alias{scPagwas_perform_regression}
\title{scPagwas_perform_regression}
\usage{
scPagwas_perform_regression(
  Pagwas,
  n.cores = 1,
  scPagwasSession = "scPagwasSession"
)
}
\arguments{
\item{Pagwas}{Pagwas data list, default is "NULL"}

\item{n.cores}{(integr)Parallel cores,default is 1. use detectCores() to check the cores in computer.}

\item{scPagwasSession}{"scPagwasSession"}
}
\value{

}
\description{
Functions for inferring relevant annotations using the polyTest model.
}
\examples{
library(scPagwas)
scPagwas_perform_regression(Pagwas, n.cores = n.cores)
}
