% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Singlecell_heritability_contributions.R
\name{scPagwas_perform_score}
\alias{scPagwas_perform_score}
\title{scPagwas_perform_score}
\usage{
scPagwas_perform_score(Pagwas, remove_outlier = TRUE)
}
\arguments{
\item{Pagwas}{pagwas}

\item{remove_outlier}{Whether to remove the outlier for scPagwas score.
storage, you can set a large number to save time.}
}
\value{
Pagwas result list including scPagwas.gPAS.score
}
\description{
Get the scPagwas score for each cells
}
