% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scPagwas_perform_score.R
\name{scPagwas_perform_score}
\alias{scPagwas_perform_score}
\title{scPagwas_perform_score}
\usage{
scPagwas_perform_score(
  Pagwas,
  n.cores = n.cores,
  split_n = NULL,
  remove_outlier = TRUE
)
}
\arguments{
\item{Pagwas}{Pagwas data list}

\item{n.cores}{cores}

\item{split_n}{(integr)default NULL, When the cell number is too big, there may have memory errors, set split_n=10000 or other number can split the cell data.}

\item{remove_outlier}{Whether to remove the outlier for scPagwas score.}
}
\value{

}
\description{
Get the scPagwas score for each cells
}
\examples{
library(scPagwas)
scPagwas_perform_score(Pagwas)
}
