% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Singlecell_heritability_contributions.R
\name{scPagwas_score_filter}
\alias{scPagwas_score_filter}
\title{scPagwas_score_filter}
\usage{
scPagwas_score_filter(scPagwas_score)
}
\arguments{
\item{scPagwas_score}{(data.frame)}
}
\description{
filter the cPagwas_score for outliers.
}
