% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Singlecell_heritability_contributions.R
\name{scParameter_regression}
\alias{scParameter_regression}
\title{scParameter_regression}
\usage{
scParameter_regression(Pagwas_x, Pagwas_y, noise_per_snp)
}
\arguments{
\item{Pagwas_x}{x parameter for lm}

\item{Pagwas_y}{y parameter for lm}

\item{noise_per_snp}{noise}

\item{ncores}{cores for regression}
}
\description{
Find parameter estimates for the data.
}
