% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scPagwas_perform_score.R
\name{scParameter_regression}
\alias{scParameter_regression}
\title{scParameter_regression}
\usage{
scParameter_regression(Pagwas_x, Pagwas_y, noise_per_snp, n.cores = 1)
}
\arguments{
\item{Pagwas_x}{x parameter for lm}

\item{Pagwas_y}{y parameter for lm}

\item{noise_per_snp}{noise}

\item{n.cores}{cores}
}
\value{

}
\description{
Find parameter estimates for the data.
}
