% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scPagwas-data.R
\docType{data}
\name{scRNAexample}
\alias{scRNAexample}
\title{scRNAexample
A single cell example data from \code{
liver_data<-readRDS("/share/pub/qiuf/COVID/liver/Rpoly.rds")
#set Idents
liver_data@meta.data$annotation <- liver_data@meta.data$anno
Idents(liver_data) <- liver_data@meta.data$anno
#normalization
liver_data <- NormalizeData(liver_data, normalization.method = "LogNormalize", scale.factor = 10000)
#correct the cell names, need no specific signal
anno <- liver_data@meta.data$annotation
 anno<-str_replace_all(anno,"-",".")
 anno<-str_replace_all(anno," ",".")
anno[which(anno=="Inflammatory.monocyte/macrophages")]<-"Inflammatory.monocyte.macrophages"
Celltype_anno$annotation<-str_replace_all(anno,"\\+",".")
Idents(liver_data) <- as.factor(anno)
liver_data<-subset(liver_data,idents = c("Stellate.cell","unknow","γδ.T.cell"), invert = TRUE)
liver_data2<-liver_data[,sample(ncol(liver_data),1000)]
save(scRNAexample,file="/share/pub/dengcy/GWAS_Multiomics/pagwas/data/scRNAexample.RData")
}}
\format{
Seruat
}
\source{
Generated from PBC
}
\description{
scRNAexample
A single cell example data from \code{
liver_data<-readRDS("/share/pub/qiuf/COVID/liver/Rpoly.rds")
#set Idents
liver_data@meta.data$annotation <- liver_data@meta.data$anno
Idents(liver_data) <- liver_data@meta.data$anno
#normalization
liver_data <- NormalizeData(liver_data, normalization.method = "LogNormalize", scale.factor = 10000)
#correct the cell names, need no specific signal
anno <- liver_data@meta.data$annotation
 anno<-str_replace_all(anno,"-",".")
 anno<-str_replace_all(anno," ",".")
anno[which(anno=="Inflammatory.monocyte/macrophages")]<-"Inflammatory.monocyte.macrophages"
Celltype_anno$annotation<-str_replace_all(anno,"\\+",".")
Idents(liver_data) <- as.factor(anno)
liver_data<-subset(liver_data,idents = c("Stellate.cell","unknow","γδ.T.cell"), invert = TRUE)
liver_data2<-liver_data[,sample(ncol(liver_data),1000)]
save(scRNAexample,file="/share/pub/dengcy/GWAS_Multiomics/pagwas/data/scRNAexample.RData")
}
}
\examples{
data(scRNAexample)
str(scRNAexample)
}
