% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Single_data_input.R
\name{Single_data_input}
\alias{Single_data_input}
\title{Single_data_input}
\usage{
Single_data_input(
  Pagwas,
  Single_data,
  nfeatures = NULL,
  Pathway_list = NULL,
  assay = "RNA",
  min_clustercells = 5
)
}
\arguments{
\item{Pagwas}{Pagwas format}

\item{Single_data}{Input the Single data in seruat format, Idents should
be the celltypes annotation. mainly used the Single_data@assays$RNA@data}

\item{nfeatures}{The parameter for FindVariableFeatures,
NULL means select all genes}

\item{Pathway_list}{(list,character) pathway gene sets list}

\item{assay}{assay data of your single cell data to use,default is "RNA".}

\item{min_clustercells}{Threshold for total cells fo each cluster.}
}
\value{
Pagwas list including:
"Celltype_anno"    "data_mat"         "VariableFeatures" "merge_scexpr"
}
\description{
Input the Single data in seruat format
}
\examples{
library(scPagwas)
Pagwas <- list()
# Start to read the single cell data
Single_data <- readRDS(system.file("extdata", "scRNAexample.rds",
  package = "scPagwas"
))
Pagwas <- Single_data_input(
  Pagwas = Pagwas,
  assay = "RNA",
  Single_data = Single_data,
  Pathway_list = Genes_by_pathway_kegg
)
}
\author{
Chunyu Deng
}
\keyword{Single_data_input,}
\keyword{cell}
\keyword{extract}
\keyword{matrix.}
\keyword{single}
\keyword{the}
