% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Single_data_input.R
\name{Single_data_input}
\alias{Single_data_input}
\title{Single_data_input}
\usage{
Single_data_input(
  Pagwas,
  Single_data,
  nfeatures = NULL,
  Pathway_list = NULL,
  assay = "RNA",
  min_clustercells = 5
)
}
\arguments{
\item{Pagwas}{Pagwas format}

\item{Single_data}{Input the Single data in seruat format, Idents should be
the celltypes annotation. mainly used the Single_data@assays$RNA@data}

\item{nfeatures}{The parameter for FindVariableFeatures,
NULL means select all genes}

\item{Pathway_list}{(list,character) pathway gene sets list}

\item{min_clustercells}{Threshold for total cells fo each cluster.}
}
\value{
Pagwas
}
\description{
Input the Single data in seruat format
}
\examples{
library(scPagwas)
data(scRNAexample)
# Pagwas is better getting from GWAS_summary_input(),or NULL is right,
Pagwas <- Single_data_input(Pagwas = Pagwas, Single_data = scRNAexample)
}
