% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Singlecell_heritability_contributions.R
\name{Singlecell_heritability_contributions}
\alias{Singlecell_heritability_contributions}
\title{Singlecell_heritability_contributions}
\usage{
Singlecell_heritability_contributions(
  Pagwas,
  iters = 200,
  n.cores = 1,
  part = 0.5,
  SCregression = FALSE
)
}
\arguments{
\item{Pagwas}{Pagwas format, deault is NULL.}

\item{iters}{number of bootstrap iterations to perform}

\item{n.cores}{Parallel cores,default is 1. use detectCores() to check the cores in computer.}

\item{part}{number of bootstrap iterations to perform,default is 0.5}

\item{SCregression}{whether to run regression funcitons, which takes a lot of time and memmory.}
}
\value{

}
\description{
Singlecell_heritability_contributions
}
