% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Singlecell_heritability_contributions.R
\name{Singlecell_heritability_contributions}
\alias{Singlecell_heritability_contributions}
\title{Singlecell_heritability_contributions}
\usage{
Singlecell_heritability_contributions(
  Pagwas,
  n.cores = 1,
  part = 0.5,
  remove_outlier = TRUE,
  SimpleResult = TRUE
)
}
\arguments{
\item{Pagwas}{Pagwas format, deault is NULL.}

\item{n.cores}{Parallel cores,default is 1. use detectCores() to check the cores in computer.}

\item{part}{number of bootstrap iterations to perform,default is 0.5}

\item{SimpleResult}{TRUE}
}
\value{

}
\description{
Singlecell_heritability_contributions
}
